==> scripts/14_Ensemble.Rmd <==
---
title: "14_Ensemble"
author: "Eddie Pérochon"
date: "05/05/2021"
output: html_document
---

##### Script 14:  Models Ensemble #####

###################################################
#      Authors: Eddie Pérochon & Maël Doré        #
#      Contact: eddie.perochon@hotmail.com        #
###################################################

### Goals = 
# Generate stacks of all submodels for each unit
# Compute EM: median, incertitude (IQR), binary, and committee averaging
# Update boxplot of evaluations with only the selected submodels, and the EM evaluation


### Inputs 
   # Summary table for unit models (list_occ)
   # Submodel maps from Script 9
   # List of selected models for ensemble from Script 12
   # Optimized thresholds for binarization from Script 11

###

### Outputs
   # Multiple pages PDF with all selected submodel maps per unit
   # Ensemble maps of each unit: median, incertitude (IQR), binary, and committee averaging
   # Thresholds and optimized evaluation metric for Ensemble
   # Updated boxplots of evaluation metrics per unit, and globally, with only the selected submodels, and the EM evaluation
###



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# Effacer l'environnement
rm(list = ls())

# Load libraries

library(raster)
library(biomod2)
library(tidyverse)
library(ggplot2)

```


```{r}
# Load Summary table for OMU models
load(file = paste0("../input_data/occurences_df/list_occ_after_modeling.RData"))
modeled_unit <- list_occ[!is.na(list_occ$Model_ID), ]

# Load summary table for sub_models
list.submodels <- readRDS(file = "../input_data/occurences_df/list.submodels.rds")

# Choose the eval metric for selection
valid_submodels_names_Jaccard <- list.submodels$model_name[list.submodels$Ensemble_OK_Jaccard] # Jaccard
valid_submodels_names_TSS <- list.submodels$model_name[list.submodels$Ensemble_OK_TSS] # TSS

# Load the cut-off data
all.units_cutoffs <- readRDS(file = "../controls/evaluations/all.units_cutoffs.rds")

# load eval data to update boxplots with EM metric and only selected submodels
gg_jaccard <- readRDS(file = "../controls/evaluations/Full_df_Jaccard.rds")
gg_TSS <- readRDS(file = "../controls/evaluations/Full_df_TSS.rds")

```

```{r}
# Set up directory
initial.wd <- getwd()
setwd(paste0(initial.wd,"/models/"))

# Create object to keep track of cutoff research
jaccard.test.list_EM <- TSS.test.list_EM <- NULL

############## Modif between scripts ######################
 index_model_to_compute <- c(1:nrow(modeled_unit))
# index_model_to_compute <- c(1:200)
# index_model_to_compute <- c(201:400)
# index_model_to_compute <- c(401:563)
###########################################################

### Loop for all modeled unit
# for (i in 1:nrow(modeled_unit)) 
p<-progress_estimated(nrow(modeled_unit), min_time=10)

for (i in index_model_to_compute) 
{
  # i <- 1
  p$pause(0.1)$tick()$print()
  unit <- as.character(modeled_unit$biomodTag[i])
  sp <- as.character(modeled_unit$Sp_ID[i])
  
  cat(paste0("\n", Sys.time()," ------ Starts for ", unit, " = Unit N°",i,"/",nrow(modeled_unit)," ------\n"))
  
  # Create directory to store stacks by OMU
  if(!dir.exists(paste0("../outputs/By_unit/",unit))) { # Test if the folder exists already or not
    dir.create(paste0("../outputs/By_unit/",unit), recursive = T) # Create folder if absent
  }
  
  # Create directory to store stacks by species
  if(!dir.exists(paste0("../outputs/By_Species/",sp))) { # Test if the folder exists already or not
    dir.create(paste0("../outputs/By_Species/",sp), recursive = T) # Create folder if absent
  }
  
  # Create directory to store maps by OMU
  if(!dir.exists(paste0("../Maps/By_unit/",unit))) { # Test if the folder exists already or not
    dir.create(paste0("../Maps/By_unit/",unit), recursive = T) # Create folder if absent
  }
  
  # Create directory to store maps by species
  if(!dir.exists(paste0("../Maps/By_Species/",sp))) { # Test if the folder exists already or not
    dir.create(paste0("../Maps/By_Species/",sp), recursive = T) # Create folder if absent
  }
  
  ### 1/ Generate stacks of selected submodels ####
  
  ### 1.1/ Continous stacks ####
  
  # Load stacks of all submodels projections
  full_stack_cont <- readRDS(file = paste0(initial.wd,"/models/",unit,"/proj_Current/proj_Current_",unit,".rds"))
  
  Select_Jaccard_stack_cont <- full_stack_cont[[which(names(full_stack_cont) %in% valid_submodels_names_Jaccard)]]/1000
  Select_TSS_stack_cont <- full_stack_cont[[which(names(full_stack_cont) %in% valid_submodels_names_TSS)]]/1000
  
  # Save the .RData/rds
  save(Select_Jaccard_stack_cont, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont.RData"), version = "2")
  saveRDS(Select_Jaccard_stack_cont, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont.rds"), version = "2")
  save(Select_TSS_stack_cont, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont.RData"), version = "2")
  saveRDS(Select_TSS_stack_cont, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont.rds"), version = "2")
  
  # Get the occurrences coordinates
  formated.input.data <- readRDS(file = paste0(initial.wd,"/models/",unit,"/formated.input.data.rds"))
  coor_presences <- formated.input.data@coord[which(formated.input.data@data.species == 1), ]
  
  # Plot the maps for Jaccard in a multiple pages pdf
  pdf(file = paste0("../Maps/By_unit/",unit,"/Select_Jaccard_stack_cont_",unit,".pdf"), height = 8, width = 8)
  for (j in names(Select_Jaccard_stack_cont))
  {
    plot(Select_Jaccard_stack_cont[[j]])
    points(coor_presences, pch = 16)
    points(coor_presences, pch = 16, cex = 0.6, col = "red")
    title(main = j)
  }
  dev.off()
  # Copy in Species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/Select_TSS_stack_cont_",unit,".pdf"), to = paste0("../Maps/By_Species/",sp,"/Select_TSS_stack_cont_",unit,".pdf"), overwrite = T)
  
  # Plot the maps for TSS in a multiple pages pdf
  pdf(file = paste0("../Maps/By_unit/",unit,"/Select_TSS_stack_cont_",unit,".pdf"), height = 8, width = 8)
  for (j in names(Select_TSS_stack_cont))
  {
    plot(Select_TSS_stack_cont[[j]])
    points(coor_presences, pch = 16, add.plot = T)
    points(coor_presences, pch = 16, cex = 0.6, col = "red", add.plot = T)
    title(main = j)
  }
  dev.off()
  # Copy in Species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/Select_TSS_stack_cont_",unit,".pdf"), to = paste0("../Maps/By_Species/",sp,"/Select_TSS_stack_cont_",unit,".pdf"), overwrite = T)
  
  ### 1.2/ Binary stacks ####
  
  unit_cutoffs <- all.units_cutoffs[unit, , , , ]
  
  # For Jaccard selection
  Select_Jaccard_stack_bin <- Select_Jaccard_stack_cont
  for (j in names(Select_Jaccard_stack_cont)) 
  {
    pa <- strsplit(j, split = "_")[[1]][2]
    cv <- strsplit(j, split = "_")[[1]][3]
    algo <- strsplit(j, split = "_")[[1]][4]
    
    # Extract threshold
    cutoff <- unit_cutoffs[algo, cv, pa, "Jaccard"]
    
    # Generate binary map and store it in the stack
    Select_Jaccard_stack_bin[[j]] <- Select_Jaccard_stack_cont[[j]] > (cutoff/1000)
  }
  
  # For TSS selection
  Select_TSS_stack_bin <- Select_TSS_stack_cont
  for (j in names(Select_TSS_stack_cont)) 
  {
    pa <- strsplit(j, split = "_")[[1]][2]
    cv <- strsplit(j, split = "_")[[1]][3]
    algo <- strsplit(j, split = "_")[[1]][4]
    
    # Extract threshold
    cutoff <- unit_cutoffs[algo, cv, pa, "TSS"]
    
    # Generate binary map and store it in the stack
    Select_TSS_stack_bin[[j]] <- Select_TSS_stack_cont[[j]] > (cutoff/1000)
  }
  
  # Save the .RData/rds
  save(Select_Jaccard_stack_bin, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin.RData"), version = "2")
  saveRDS(Select_Jaccard_stack_bin, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin.rds"), version = "2")
  save(Select_TSS_stack_bin, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin.RData"), version = "2")
  saveRDS(Select_TSS_stack_bin, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin.rds"), version = "2")
  
  # Plot the maps for Jaccard in a multiple pages pdf
  pdf(file = paste0("../Maps/By_unit/",unit,"/Select_Jaccard_stack_bin_",unit,".pdf"), height = 8, width = 8)
  for (j in names(Select_Jaccard_stack_bin))
  {
    plot(Select_Jaccard_stack_bin[[j]])
    points(coor_presences, pch = 16, add.plot = T)
    points(coor_presences, pch = 16, cex = 0.6, col = "red", add.plot = T)
    title(main = j)
  }
  dev.off()
  # Copy in Species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/Select_TSS_stack_bin_",unit,".pdf"), to = paste0("../Maps/By_Species/",sp,"/Select_TSS_stack_bin_",unit,".pdf"), overwrite = T)
  
  # Plot the maps for TSS in a multiple pages pdf
  pdf(file = paste0("../Maps/By_unit/",unit,"/Select_TSS_stack_bin_",unit,".pdf"), height = 8, width = 8)
  for (j in names(Select_TSS_stack_bin))
  {
    plot(Select_TSS_stack_bin[[j]])
    points(coor_presences, pch = 16, add.plot = T)
    points(coor_presences, pch = 16, cex = 0.6, col = "red", add.plot = T)
    title(main = j)
  }
  dev.off()
  # Copy in Species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/Select_TSS_stack_bin_",unit,".pdf"), to = paste0("../../maps/By_Species/",sp,"/Select_TSS_stack_bin_",unit,".pdf"), overwrite = T)
  
  
  ### 2/ Ensemble
  
  cat(paste0("\n", Sys.time()," ------ Ensemble computation for ", unit, " = Unit N°",i,"/",nrow(modeled_unit)," ------\n"))
  
  ### 2.1/ Continuous Ensemble ####
  
  Ensemble_Jaccard_median <- calc(x = Select_Jaccard_stack_cont, fun = median, na.rm = T)
  Ensemble_TSS_median <- calc(x = Select_TSS_stack_cont, fun = median, na.rm = T)
  
  save(Ensemble_Jaccard_median, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_median.RData"), version = "2")
  saveRDS(Ensemble_Jaccard_median, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_median.rds"), version = "2")
  save(Ensemble_TSS_median, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_median.RData"), version = "2")
  saveRDS(Ensemble_TSS_median, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_median.rds"), version = "2")
  
  ### 2.2/ Incertitude on Ensemble (IQR) ####
  
  Ensemble_Jaccard_IQR <- calc(x = Select_Jaccard_stack_cont, fun = IQR, na.rm = T)
  Ensemble_TSS_IQR <- calc(x = Select_TSS_stack_cont, fun = IQR, na.rm = T)
  
  save(Ensemble_Jaccard_IQR, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_IQR.RData"), version = "2")
  saveRDS(Ensemble_Jaccard_IQR, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_IQR.rds"), version = "2")
  save(Ensemble_TSS_IQR, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_IQR.RData"), version = "2")
  saveRDS(Ensemble_TSS_IQR, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_IQR.rds"), version = "2")
  
  ### 2.3/ Committee Averaging for binary map ####
  
  # Jaccard
  Ensemble_Jaccard_CA <- sum(Select_Jaccard_stack_bin)                                        # Sum all binary maps to get the nb of votes
  Ensemble_Jaccard_CA[Ensemble_Jaccard_CA < (nlayers(Select_Jaccard_stack_bin) / 2)] <- 0     # If less then half votes for presence, ensemble displays an absence
  Ensemble_Jaccard_CA[Ensemble_Jaccard_CA >= (nlayers(Select_Jaccard_stack_bin) / 2)] <- 1    # If half or more votes for presence, ensemble displays a presence
  # plot(Ensemble_Jaccard_CA)
  
  # TSS
  Ensemble_TSS_CA <- sum(Select_TSS_stack_bin)                                    # Sum all binary maps to get the nb of votes
  Ensemble_TSS_CA[Ensemble_TSS_CA < (nlayers(Select_TSS_stack_bin) / 2)] <- 0     # If less then half votes for presence, ensemble displays an absence
  Ensemble_TSS_CA[Ensemble_TSS_CA >= (nlayers(Select_TSS_stack_bin) / 2)] <- 1    # If half or more votes for presence, ensemble displays a presence
  # plot(Ensemble_TSS_CA)
  
  save(Ensemble_Jaccard_CA, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_CA.RData"), version = "2")
  saveRDS(Ensemble_Jaccard_CA, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_CA.rds"), version = "2")
  save(Ensemble_TSS_CA, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_CA.RData"), version = "2")
  saveRDS(Ensemble_TSS_CA, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_CA.rds"), version = "2")
  
  ### 2.4/ Binary map based on evaluation of all observations fitted in all selected models ####
  
  # Note this is biased because evaluation is done on observations used to calibrate some of the models, and nb of PsA depend on model type (more PsA for "restricted")
  
  # Retrieve input data
  obs.data <- formated.input.data@data.species
  obs.data[is.na(obs.data)] <- 0   # Change pseudo-absences (NA) in real absences (0) to be able to compute confusion matrix
  
  # Retrieve coordinates
  coor_all <- formated.input.data@coord
  
  ## Jaccard
  
  # Extract predictions from the EM
  pred.data <- raster::extract(Ensemble_Jaccard_median*1000, coor_all)
  
  jaccard.test <- NULL   # Object to store the cutoff research                                
  for(cutoff in seq(0, 1000, by = 1))                    
  {
    pred.pa <- pred.data              # Predictions of the model for the test set
    pred.pa[pred.pa < cutoff] <- 0    # We assigne a 0 (Absence) for all predicts under the cutoff 
    pred.pa[pred.pa >= cutoff] <- 1   # We assign a 1 (presence) for all predicts equal or above the cutoff
    
    # We compare predicts and inputs = confusion matrix
    TP <- sum(obs.data == 1 & pred.pa == 1, na.rm = T)   # True positives
    FN <- sum(obs.data == 1 & pred.pa == 0, na.rm = T)   # False negatives
    FP <- sum(obs.data == 0 & pred.pa == 1, na.rm = T)   # False positives
    
    # Compute Jaccard index
    jaccard <- TP / (TP + FP + FN)
    
    # Store all infos regarding the cut.off test for TSS in a df
    jaccard.test <- rbind.data.frame(jaccard.test,
                                     data.frame(cutoff = cutoff,
                                                TP = TP,
                                                FN = FN,
                                                FP = FP,
                                                jaccard = jaccard))
  }
  # Store the test track
  jaccard.test.list_EM[[unit]] <- jaccard.test
  
  # Extract the cutoff as the value that maximize the evaluation
  EM.cutoff_Jaccard <- round(mean(jaccard.test$cutoff[which(jaccard.test$jaccard == max(jaccard.test$jaccard))])) # Take the mean value because several cutoffs can maximize the evaluation metric
  EM.metric_Jaccard <- round(jaccard.test$jaccard[round(which(jaccard.test$cutoff == EM.cutoff_Jaccard))],3)
  # Store it in list_occ
  list_occ$EM.cutoff_Jaccard[list_occ$biomodTag == unit] <- EM.cutoff_Jaccard
  list_occ$EM.metric_Jaccard[list_occ$biomodTag == unit] <- EM.metric_Jaccard
  
  # Binariazation of the continous EM
  Ensemble_Jaccard_bin <- BinaryTransformation(Ensemble_Jaccard_median, EM.cutoff_Jaccard/1000)
  # plot(Ensemble_Jaccard_bin)
  
  ## TSS
  
  # Extract predictions from the EM
  pred.data <- raster::extract(Ensemble_TSS_median*1000, coor_all)
  
  TSS.test <- NULL   # Object to store the cutoff research                                
  
  for(cutoff in seq(0, 1000, by = 1))                    
  {
    pred.pa <- pred.data              # Predictions of the model for the test set
    pred.pa[pred.pa < cutoff] <- 0    # We assigne a 0 (Absence) for all predicts under the cutoff 
    pred.pa[pred.pa >= cutoff] <- 1   # We assign a 1 (presence) for all predicts equal or above the cutoff
    
    # We compare predicts and inputs = confusion matrix
    TP <- sum(obs.data == 1 & pred.pa == 1, na.rm = T)   # True positives
    FN <- sum(obs.data == 1 & pred.pa == 0, na.rm = T)   # False negatives
    FP <- sum(obs.data == 0 & pred.pa == 1, na.rm = T)   # False positives
    TN <- sum(obs.data == 0 & pred.pa == 0, na.rm = T)   # True negatives
    
    sensitivity <- round(TP / (TP + FN), 3) # Ability to detect presence among presences points
    specificity <- round(TN / (TN + FP), 3)  # Ability to detect absences among absences points
    
    TSS <- sensitivity + specificity - 1  # Compute TSS
    
    # Store all infos regarding the cut.off test for TSS in a df
    TSS.test <- rbind.data.frame(TSS.test,                  
                                 data.frame(cutoff = cutoff,
                                            sensitivity = sensitivity,
                                            specificity = specificity,
                                            TSS = TSS))
  }
  # Store the test track
  TSS.test.list_EM[[unit]] <- TSS.test
  
  # Extract the cutoff as the value that maximize the evaluation
  EM.cutoff_TSS <- round(mean(TSS.test$cutoff[which(TSS.test$TSS == max(TSS.test$TSS))])) # Take the mean value because several cutoffs can maximize the evaluation metric
  EM.metric_TSS <- round(TSS.test$TSS[round(which(TSS.test$cutoff == EM.cutoff_TSS))],3)
  # Store it in list_occ
  list_occ$EM.cutoff_TSS[list_occ$biomodTag == unit] <- EM.cutoff_TSS
  list_occ$EM.metric_TSS[list_occ$biomodTag == unit] <- EM.metric_TSS
  
  # Binariazation of the continous EM
  Ensemble_TSS_bin <- BinaryTransformation(Ensemble_TSS_median, EM.cutoff_TSS/1000)
  # plot(Ensemble_Jaccard_bin)
  
  ## Save binary EMs
  save(Ensemble_Jaccard_bin, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_bin.RData"), version = "2")
  saveRDS(Ensemble_Jaccard_bin, file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_bin.rds"), version = "2")
  save(Ensemble_TSS_bin, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_bin.RData"), version = "2")
  saveRDS(Ensemble_TSS_bin, file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_bin.rds"), version = "2")
  
  ### 2.5/ Plot all EMs on one page ####
  
  # Jaccard
  all_EM_Jaccard <- stack(Ensemble_Jaccard_median, Ensemble_Jaccard_IQR, Ensemble_Jaccard_bin, Ensemble_Jaccard_CA)
  names(all_EM_Jaccard) <- c("Continuous", "Incertitude (IQR)", "Binary", "Committee Averaging")
  
  pdf(file = paste0("../Maps/By_unit/",unit,"/Ensemble_maps_Jaccard_",unit,".pdf"), height = 6, width = 7)
  internal_margins <- par()$mar
  external_margins <- par()$oma
  par(mar = c(5.1,6.1,9.1,3.1))
  par(oma = c(1,2,4,3))
  plot(all_EM_Jaccard)
  title(main = unit, outer = T)
  par(mar = internal_margins, oma = external_margins)
  dev.off()
  # Copy in species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/Ensemble_maps_Jaccard_",unit,".pdf"), to = paste0("../Maps/By_Species/",sp,"/Ensemble_maps_Jaccard_",unit,".pdf"), overwrite = T)
  
  save(all_EM_Jaccard, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard.RData"), version = "2")
  saveRDS(all_EM_Jaccard, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard.rds"), version = "2")
  
  # TSS
  all_EM_TSS <- stack(Ensemble_TSS_median, Ensemble_TSS_IQR, Ensemble_TSS_bin, Ensemble_TSS_CA)
  names(all_EM_TSS) <- c("Continuous", "Incertitude (IQR)", "Binary", "Committee Averaging")
  
  pdf(file = paste0("../Maps/By_unit/",unit,"/Ensemble_maps_TSS_",unit,".pdf"), height = 6, width = 7)
  internal_margins <- par()$mar
  external_margins <- par()$oma
  par(mar = c(5.1,6.1,9.1,3.1))
  par(oma = c(1,2,4,3))
  plot(all_EM_TSS)
  title(main = unit, outer = T)
  par(mar = internal_margins, oma = external_margins)
  dev.off()
  # Copy in species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../%aps/By_unit/",unit,"/Ensemble_maps_TSS_",unit,".pdf"), to = paste0("../Maps/By_Species/",sp,"/Ensemble_maps_TSS_",unit,".pdf"), overwrite = T)
  
  save(all_EM_TSS, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS.RData"), version = "2")
  saveRDS(all_EM_TSS, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS.rds"), version = "2")
  
  ### 3/ Update boxplots for evaluation with only selected submodels and Eval metric value ####
  
  cat(paste0("\n", Sys.time()," ------ Evaluation boxplots for ", unit, " = Unit N°",i,"/",nrow(modeled_unit)," ------\n"))
  
  unit_model_type <- modeled_unit$initial_model_type[i]
  
  ## Jaccard
  
  gg_jaccard_unit <- gg_jaccard[gg_jaccard$unit == unit, ]
  
  # Update the df
  gg_jaccard_unit <- gg_jaccard_unit %>% 
    mutate(model_name = paste0(unit, "_", PA_sets, "_", CV_runs, "_", Algo)) %>%  # Add a model_name column
    filter(model_name %in% valid_submodels_names_Jaccard) %>%  # Filter only selected submodels
    add_row(unit = unit, Algo = "EM", value = EM.metric_Jaccard) # Add the EM value
  
  # Order the factor levels
  gg_jaccard_unit$Algo <- factor(gg_jaccard_unit$Algo, levels = c("RF", "GBM", "ANN", "EM"))
  
  # Plot
  pdf(file = paste0("../controls/evaluations/By_unit/boxplots/Jaccard/Postselect_Eval_Jaccard_boxplot_",unit,".pdf"), height = 6, width = 10)
  g <- ggplot(gg_jaccard_unit, aes(x = Algo, y = value, fill = Algo)) +
    geom_boxplot(show.legend = F) +
    ylim(c(min(gg_jaccard_unit$value)*0.9, 1)) +
    geom_hline(yintercept = c(0.6, 0.95)[c("complete", "restricted") == unit_model_type], col = "red", lwd = 1, lty = 2) +
    labs(title = paste0("Evaluation of selected models for ",unit," with Jaccard indices"),
         subtitle = paste0("Model type = ",unit_model_type),
         y = "Jaccard indices",
         x = "Algorithm or Ensemble")
  print(g)
  dev.off()
  
  ## TSS
  
  gg_TSS_unit <- gg_TSS[gg_TSS$unit == unit, ]
  
  # Update the df
  gg_TSS_unit <- gg_TSS_unit %>% 
    mutate(model_name = paste0(unit, "_", PA_sets, "_", CV_runs, "_", Algo)) %>%  # Add a model_name column
    filter(model_name %in% valid_submodels_names_TSS) %>%  # Filter only selected submodels
    add_row(unit = unit, Algo = "EM", value = EM.metric_TSS) # Add the EM value
  
  # Order the factor levels
  gg_TSS_unit$Algo <- factor(gg_TSS_unit$Algo, levels = c("RF", "GBM", "ANN", "EM"))
  
  # Plot
  pdf(file = paste0("../controls/evaluations/By_unit/boxplots/TSS/Postselect_Eval_TSS_boxplot_",unit,".pdf"), height = 6, width = 10)
  g <- ggplot(gg_TSS_unit, aes(x = Algo, y = value, fill = Algo)) +
    geom_boxplot(show.legend = F) +
    ylim(c(min(gg_TSS_unit$value)*0.9, 1)) +
    geom_hline(yintercept = c(0.6, 0.95)[c("complete", "restricted") == unit_model_type], col = "red", lwd = 1, lty = 2) +
    labs(title = paste0("Evaluation of selected models for ",unit," with TSS indices"),
         subtitle = paste0("Model type = ",unit_model_type),
         y = "TSS indices",
         x = "Algorithm or Ensemble")
  print(g)
  dev.off()
 
  cat(paste0("\n", Sys.time()," ------ Done for ", unit, " = Unit N°",i,"/",nrow(modeled_unit)," ------\n"))
  
}

```
