==> scripts/15_Clipping.Rmd <==
---
title: "15_Clipping"
author: "Eddie Pérochon"
date: "07/05/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


##### Script 15: Clipping maps#####

###################################################
#      Authors: Eddie Pérochon & Maël Doré        #
#      Contact: eddie.perochon@hotmail.com        #
###################################################


### Goals = 
# Use alpha-hull around occurrences points and a buffer proportional to the quality of presence sampling to clip the maps
# Clip EM maps and individual submodel projections, for each unit
 
### Inputs 
   # Buffer for continental borders
   # Buffer for Andean regions
   # unit occurrence coordinates (to draw the alpha-hull) from Script 05
   # Stack of EM for each unit from script 14
   # Stack of submodels for each unit from script 14
###

### Outputs
  
   # Buffer value based on 80%/95% quantile of occurrences distance to closest point
   # Alpha-hull for each unit
   # Plots of final clipped EM maps

###

```{r}
# Effacer l'environnement
rm(list = ls())

# Load libraries
library(raster)
library(rgeos)
library(rgdal)
library(alphahull)
library(geosphere)
library(spatialEco)
library(rnaturalearth)
library(dplyr)
library(rnaturalearthdata)
library(rnaturalearthhires)
library(sp)
library(sf)

```

```{r}
My_crop<-function(x){
xmin = -133 ; xmax = -30 ; ymin = -37 ; ymax = 48
e <- extent(c(xmin,xmax,ymin,ymax))

cropped<-crop(x, e)

return(cropped)
}
```

```{r}
#continent mask creation
continent_mask<-rnaturalearthhires::countries10
continent_mask<-My_crop(continent_mask)

continent_mask<-st_as_sf(continent_mask) 
sf_continent_mask <- st_union(continent_mask)
continent_mask <- as(sf_continent_mask, "Spatial") # If you want sp

continent_mask_sp<- spTransform(continent_mask,
             CRSobj=CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))

load(file="../input_data/envData/Envstack.RData") #for rastering


  continent_mask <- rasterize(continent_mask_sp, Var_stack_final, field=1)
  
continent_mask@data@values[continent_mask@data@values==0]<-NA
continent_mask@data@values[continent_mask@data@values==1]<-0
  
save(continent_mask, file="../input_data/envData/continent_mask.RData")
saveRDS(continent_mask, file="../input_data/envData/continent_mask.RDS") 

```


```{r}
# Load Summary table for OMU models
load(file = paste0("../input_data/occurences_df/list_occ_after_modeling.RData"))
modeled_unit <- list_occ[!is.na(list_occ$Model_ID), ]


# Load environmental stack to use as reference for extent and CRS
envData <- readRDS(file = paste0("../input_data/envData/Envstack.rds"))

# Load mask for continent borders
continent_mask <- readRDS(file = paste0("../input_data/envData/continent_mask.rds")) #already in MW so no need to change proj

# Load buffers for Andes
Andes_East <- spTransform( readOGR(dsn = paste0("../input_data//biogeo_masks"), layer = "neo_e"),
                           CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))


  ####Andes_west, commented bc too long###########

# Andes_West <- spTransform( readOGR(dsn = paste0("../input_data/biogeo_masks"), layer = "neo_w"),
#                            CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
# 
# CN<-spTransform(readOGR(dsn = paste0("../input_data//biogeo_masks"), layer = "cold_n"),
# CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
# CA<-spTransform(readOGR(dsn = paste0("../input_data//biogeo_masks"), layer = "caribbean"),
# CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
# TN<-spTransform(readOGR(dsn = paste0("../input_data//biogeo_masks"), layer = "temp_n"),
# CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
# 
# Andes_West<-st_as_sf(Andes_West) #We want North America and caribbean on this layer too because it will crop points that are in this region if we don't do it
# TN<-st_as_sf(TN)#USA
# CN<-st_as_sf(CN)#Canada & NorthUSA
# CA<-st_as_sf(CA)#Caribbean
# 
# Andes_West<-rbind(Andes_West,TN) #Merge all polygons
# Andes_West<-rbind(Andes_West,CN)
# Andes_West<-rbind(Andes_West,CA)
# 
# 
# sf_andwest_mask <- st_union(Andes_West) #keep one polygon
# Andes_West<- as(sf_andwest_mask, "Spatial") #get back to sp
# save(Andes_West, file="../input_data/Biogeo_masks/Andes_West.RData")

# # Rename CRS with proper order for arguments to avoid warnings
# Andes_East@proj4string <- CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")
# Andes_West@proj4string <- CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")
load("../input_data/Biogeo_masks/Andes_West.RData")

# Rasterize the buffers
Andes_East.rasterized <- raster::rasterize(x = Andes_East, y = envData, field = 1)
Andes_West.rasterized <- raster::rasterize(x = Andes_West, y = envData, field = 1)
# plot(Andes_East.rasterized)
# plot(Andes_West.rasterized, add =T)

# Projection in Mercador (EPSG:3857) to allow intersection function to work properly
# Andes_East <-  spTransform(Andes_East , CRSobj = "+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext  +no_defs")
# Andes_West <-  spTransform(Andes_West , CRSobj = "+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext  +no_defs")

# plot(Andes_East)
# plot(Andes_West, add = T)

# Create the variable to keep track of location of OMU regarding Andes slopes
list_occ$Andes_side <-  NA 

# Functions to convert alpha-hull into spatial objects
source("../functions/Alpha_functions.R", local = TRUE) 
```


```{r}
############## Modif between scripts ######################
###for bugs
index_model_to_compute <- c(3,4,7,20,32,90,91,92,93,94,95)
                            
#index_model_to_compute <- c(1:nrow(modeled_unit))
# index_model_to_compute <- c(1:200)
# index_model_to_compute <- c(201:400)
# index_model_to_compute <- c(401:563)
###########################################################

### Loop for all unit


# Units en erreur à utiliser avec un alpha-shape plutôt que alpha-hull
#c()

### Loop for all modeled OMU/unit
p<-progress_estimated(nrow(modeled_unit), min_time=10)
for (i in index_model_to_compute) 
{
  
  p$pause(0.1)$tick()$print()
  unit <- as.character(modeled_unit$biomodTag[i])
  unit2 <- as.character(modeled_unit$code_unit[i]) #before biomod2 the tag was different
  sp <- as.character(modeled_unit$Sp_ID[i])
  
  cat(paste0("\n", Sys.time()," ------ Starts for ", unit, " = Unit N°",i,"/",nrow(modeled_unit)," ------\n"))
  
  ### Load unit occurrence points
  load(paste0("../input_data/Species_data/Spatial_Points_Objects/occurrences_", unit2,".RData")) # Load Spatial object with occurrences
  unit_points<-unit_points2
  
  unit_points2<- SpatialPointsDataFrame(unit_points2[, c("x","y")], 
                                                    unit_points2,
                                                    proj4string= CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
  
  unit_points<- spTransform( SpatialPointsDataFrame(unit_points[, c("x","y")], 
                                                    unit_points,
                                                    proj4string= CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs")),
                             CRS= CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")) #need to get unit_points back to WGS for dist calc
  
  

  # Load Ensembles (continous, ensemble, binary, committee averaging) for Jaccard indices
  Ensemble_Jaccard_median <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_median.rds"))
  Ensemble_Jaccard_bin <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_bin.rds"))
  Ensemble_Jaccard_IQR <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_IQR.rds"))
  Ensemble_Jaccard_CA <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_CA.rds"))

  # all_EM_Jaccard <- stack(Ensemble_Jaccard_median, Ensemble_Jaccard_IQR, Ensemble_Jaccard_bin, Ensemble_Jaccard_CA)
  # names(all_EM_Jaccard) <- c("Continuous", "Incertitude (sd)", "Binary", "Committee Averaging")
  # save(all_EM_Jaccard, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard.RData"), version = "2")
  # saveRDS(all_EM_Jaccard, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard.rds"), version = "2")

  # Load Ensembles (continous, ensemble, binary, committee averaging) for TSS indices
  Ensemble_TSS_median <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_median.rds"))
  Ensemble_TSS_bin <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_bin.rds"))
  Ensemble_TSS_IQR <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_IQR.rds"))
  Ensemble_TSS_CA <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_TSS_CA.rds"))

  # all_EM_TSS <- stack(Ensemble_TSS_median, Ensemble_TSS_IQR, Ensemble_TSS_bin, Ensemble_TSS_CA)
  # names(all_EM_TSS) <- c("Continuous", "Incertitude (sd)", "Binary", "Committee Averaging")
  # save(all_EM_TSS, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS.RData"), version = "2")
  # saveRDS(all_EM_TSS, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS.rds"), version = "2")

  # Load submodels stacks
  Select_Jaccard_stack_cont <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont.rds"))
  Select_TSS_stack_cont <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont.rds"))
  Select_Jaccard_stack_bin <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin.rds"))
  Select_TSS_stack_bin <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin.rds"))
  
  
  ### 1/ Compute the buffer distance as the 80% or 95% quantile the distance to the closest occurrence points among occurrence points for each OMU ####

  # To be seen as an estimation of the uncertainty surrounding the completeness of our survey.
  # Well surveyed unit should have short distances between occurrences, and no isolated points. Thus the distribution map can be clipped tighly around points
  # Unsufficiently surveyed OMU will display isolated points, thus the buffer must be wide to encompass the uncertainty around the distribution limits of this OMU
  # Use 95% quantile instead of maximum distance to avoid single isolated point to drive the size of the buffer. Allow 5% of outliers

  Coord_unit <- unit_points@coords

  Occ.dist = geosphere::distm(x = Coord_unit)/1000 # Geometric distance on WGS84 ellipsoid, in km

  # Remove diagonal (distance to thmeselves = 0 km)
  diag(Occ.dist) <- NA
  # Compute distance to closest other occurrence point
  Occ.dist_min <- apply(X = Occ.dist, MARGIN = 2, FUN = min, na.rm = T)

  # Get the 80% and 95% quantile and max.distance
  Uncertainty.80 <- round(quantile(x = Occ.dist_min, probs = 0.8),1)
  Uncertainty.95 <- round(quantile(x = Occ.dist_min, probs = 0.95),1)
  Max.dist <- round(max(Occ.dist_min),1)

  # Save in list_occ
  list_occ$Uncertainty.80[list_occ$biomodTag == unit] <- Uncertainty.80
  list_occ$Uncertainty.95[list_occ$biomodTag == unit] <- Uncertainty.95
  list_occ$Max.dist[list_occ$biomodTag == unit] <- Max.dist


  ### 2/ Generate alpha-hull buffer ####

  # Temporary projection of the spatial object of occurrences in Moll
  proj.points <-  unit_points2
  tryCatch(
   {
     Hull <- ahull_to_SPLDF(ahull(proj.points@coords, alpha = 1000)) # Try an alpha-hull first
   },
   error = function(e) # If fails, do an alpha shape instead
     {
       cat(paste0("Alpha-hull failed for ", unit, ", use alpha-shape instead \n"))
       cat("ERROR :",conditionMessage(e), "\n")   # Display the error message but do not stop the function
       Hull <<- ashape_to_SPLDF(ashape(proj.points@coords, alpha = 1000))
     })

  plot(Hull)

  # Hull <- ahull(proj.points@coords, alpha = 1000000) # alpha diameter = 1000km
  # Hull <- ahull_to_SPLDF(x = Hull)

  # For cases that cannot handle alpha-hull, use alpha-shape
  # Hull <- ashape(proj.points@coords, alpha = 1000000)
  # Hull <- ashape_to_SPLDF(x = Hull)

  # Save crs of the newly created alpha-hull
  Hull@proj4string <- proj.points@proj4string

  # Extend the shape with a buffer based on the quality of presence sampling (Uncertainty.80)
  # As a minimum, use the same buffer distance used to draw PsA = 120 km
  Buffer_80 <- gBuffer(spgeom = Hull, width = max(120, Uncertainty.80))
  Buffer_95 <- gBuffer(spgeom = Hull, width = max(120, Uncertainty.95))

  # Remove holes
  Buffer_80_unholed <- spatialEco::remove.holes(Buffer_80)
  Buffer_95_unholed <- spatialEco::remove.holes(Buffer_95)

  # Retransposition in Moll
  Buffer_80.Moll <- spTransform(Buffer_80_unholed, CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
  Buffer_95.Moll <- spTransform(Buffer_95_unholed, CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))

plot(Ensemble_Jaccard_median)
  points(unit_points2, col = "red", cex = 1, pch = 16)
   plot(Buffer_80.Moll, border = "dodgerblue", lwd = 2, col = NA, add =T)
   plot(Buffer_95.Moll, lwd = 2, add =T)

  # Rasterization
  Buffer_80.rasterized <- rasterize(x = Buffer_80.Moll, y = Ensemble_Jaccard_median, field = 1)
  Buffer_95.rasterized <- rasterize(x = Buffer_95.Moll, y = Ensemble_Jaccard_median, field = 1)
  plot(Buffer_80.rasterized, add = T)
  # plot(Buffer_95.rasterized, add = T)

  # Save final buffers
  save(Buffer_80.rasterized, file = paste0("../outputs/By_unit/",unit,"/Buffer_80.RData"), version = "2")
  saveRDS(Buffer_80.rasterized, file = paste0("../outputs/By_unit/",unit,"/Buffer_80.rds"), version = "2")
  save(Buffer_95.rasterized, file = paste0("../outputs/By_unit/",unit,"/Buffer_95.RData"), version = "2")
  saveRDS(Buffer_95.rasterized, file = paste0("../outputs/By_unit/",unit,"/Buffer_95.rds"), version = "2")

  
  ### 3/ Clip the EM and submodels maps ####
  
  ## 3.1/ Clip all EM and submodels with the buffers ####
  
  # Load buffers
  Buffer_80.rasterized <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Buffer_80.rds"))
  Buffer_95.rasterized <- readRDS(file = paste0("../outputs/By_unit/",unit,"/Buffer_95.rds"))

  ## 3.1.1/ EM maps

  # Make a stack of EMs, both Eval metrics
  full_stack <- stack(Ensemble_Jaccard_median, Ensemble_Jaccard_IQR, Ensemble_Jaccard_bin, Ensemble_Jaccard_CA, Ensemble_TSS_median, Ensemble_TSS_IQR, Ensemble_TSS_bin, Ensemble_TSS_CA)

  # Apply buffers
  full_stack_cropped_80 <- mask(full_stack, Buffer_80.rasterized)
  full_stack_cropped_95 <- mask(full_stack, Buffer_95.rasterized)

  # Make a full stack for all Eval metric and buffers
  full_stack_cropped_by_buffers <- stack(full_stack_cropped_80, full_stack_cropped_95)

  # Create names for the layers
  Ensemble <- rep("Ensemble", 16)
  Eval <- c(rep("Jaccard", 4), rep("TSS", 4))
  Type <- c("median", "IQR", "bin", "CA")
  Crop <- rep("cropped", 16)
  Buffer <- c(rep("80", 8), rep("95", 8))

  layer_names_df <- data.frame(Ensemble = Ensemble, Eval = Eval, Type = Type, Crop = Crop, Buffer = Buffer, stringsAsFactors = F)
  layer_names <- apply(X = layer_names_df, MARGIN = 1 , FUN = paste, collapse = "_")

  names(full_stack_cropped_by_buffers) <- layer_names

  ## 3.1.2/ For submodels maps
  
  # Apply buffers
  Select_Jaccard_stack_cont_cropped_80 <- mask(Select_Jaccard_stack_cont, Buffer_80.rasterized)
  Select_TSS_stack_cont_cropped_80 <- mask(Select_TSS_stack_cont, Buffer_80.rasterized)
  Select_Jaccard_stack_cont_cropped_95 <- mask(Select_Jaccard_stack_cont, Buffer_95.rasterized)
  Select_TSS_stack_cont_cropped_95 <- mask(Select_TSS_stack_cont, Buffer_95.rasterized)
  
  Select_Jaccard_stack_bin_cropped_80 <- mask(Select_Jaccard_stack_bin, Buffer_80.rasterized)
  Select_TSS_stack_bin_cropped_80 <- mask(Select_TSS_stack_bin, Buffer_80.rasterized)
  Select_Jaccard_stack_bin_cropped_95 <- mask(Select_Jaccard_stack_bin, Buffer_95.rasterized)
  Select_TSS_stack_bin_cropped_95 <- mask(Select_TSS_stack_bin, Buffer_95.rasterized)
  
  ## 3.2/ Check within which Andean slope lies the occurrences, and crop out the slope where no occurrences have been recorded ####
  
  # 3.2.1/ For EM maps

  # Checkif presence points fall into the Andes slopes (used projected object in Mercator)
  test_East <- gIntersection(spgeom1 = Andes_East, spgeom2 = proj.points, checkValidity = 2)
  test_West <- gIntersection(spgeom1 = Andes_West, spgeom2 = proj.points, checkValidity = 2)

  ### Plot
  # plot(Andes_East)
  # points(test_East, pch = 16, col ="limegreen")
  # points(test_West, pch = 16, col ="red")
  ###

  # Clip out the slope where no occurrences have been recorded
if(proj.points$y)  
  if (length(test_East)*length(test_West) == 0) { # Case with occurrences only on one side

    if (length(test_West)==0) { # Case with occurrences on the East side

      full_stack_cropped_by_buffers_and_slopes <- mask(full_stack_cropped_by_buffers, Andes_East.rasterized)

      list_occ$Andes_side[which(list_occ$biomodTag == unit)] <- "East"

    } else { # Case with occurrences on the West side

      full_stack_cropped_by_buffers_and_slopes <- mask(full_stack_cropped_by_buffers, Andes_West.rasterized)

      list_occ$Andes_side[which(list_occ$biomodTag == unit)] <- "West"

    }

  } else { # Case with occurrences on both sides = no additionnal cropping

    full_stack_cropped_by_buffers_and_slopes <- full_stack_cropped_by_buffers

    list_occ$Andes_side[which(list_occ$biomodTag == unit)] <- "Both"
  }

  ### Add 0 value for terrestrial pixels to be able to distinguish from oceans

  full_stack_cropped_by_buffers_with_borders <- full_stack_cropped_by_buffers_and_slopes
  for (j in 1:nlayers(full_stack_cropped_by_buffers))
  {
    layer_name <- names(full_stack_cropped_by_buffers)[j]

    # Extract values from layer
    temp_data <- full_stack_cropped_by_buffers_and_slopes[[j]]@data@values
    temp_data[which(is.na(temp_data))] <- 0 # Add 0 instead of NA

    temp_layer <- Ensemble_Jaccard_bin
    temp_layer@data@values <- temp_data
    names(temp_layer) <- layer_name

    # Mask with a layer of environnemental variable which have continental borders
    temp_layer <- mask(temp_layer, envData[[1]])

    # Store in new stack
    full_stack_cropped_by_buffers_with_borders[[j]] <- temp_layer

    # Save individually
    saveRDS(temp_layer, file = paste0("../outputs/By_unit/",unit,"/",layer_name,".rds"), version = "2")

  }

  # Extract stacks per Eval metric and Buffer
  all_EM_Jaccard_buffer_80 <- full_stack_cropped_by_buffers_with_borders[[1:4]]
  all_EM_TSS_buffer_80 <- full_stack_cropped_by_buffers_with_borders[[5:8]]
  all_EM_Jaccard_buffer_95 <- full_stack_cropped_by_buffers_with_borders[[9:12]]
  all_EM_TSS_buffer_95 <- full_stack_cropped_by_buffers_with_borders[[13:16]]

  # Save stacks
  save(all_EM_Jaccard_buffer_80, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard_buffer_80.RData"), version = "2")
  saveRDS(all_EM_Jaccard_buffer_80, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard_buffer_80.rds"), version = "2")
  save(all_EM_Jaccard_buffer_95, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard_buffer_95.RData"), version = "2")
  saveRDS(all_EM_Jaccard_buffer_95, file = paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard_buffer_95.rds"), version = "2")
  save(all_EM_TSS_buffer_80, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS_buffer_80.RData"), version = "2")
  saveRDS(all_EM_TSS_buffer_80, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS_buffer_80.rds"), version = "2")
  save(all_EM_TSS_buffer_95, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS_buffer_95.RData"), version = "2")
  saveRDS(all_EM_TSS_buffer_95, file = paste0("../outputs/By_unit/",unit,"/all_EM_TSS_buffer_95.rds"), version = "2")

  ## 3.2.2/ For Submodel maps
  
  # Get Andes side info
  Andes_side <- list_occ$Andes_side[which(list_occ$biomodTag == unit)]
  
  # Apply buffers
  if (Andes_side == "East") {
    
    Select_Jaccard_stack_cont_cropped_80 <- mask(Select_Jaccard_stack_cont_cropped_80, Andes_East.rasterized)
    Select_TSS_stack_cont_cropped_80 <- mask(Select_TSS_stack_cont_cropped_80, Andes_East.rasterized)
    Select_Jaccard_stack_cont_cropped_95 <- mask(Select_Jaccard_stack_cont_cropped_95, Andes_East.rasterized)
    Select_TSS_stack_cont_cropped_95 <- mask(Select_TSS_stack_cont_cropped_95, Andes_East.rasterized)
    
    Select_Jaccard_stack_bin_cropped_80 <- mask(Select_Jaccard_stack_bin_cropped_80, Andes_East.rasterized)
    Select_TSS_stack_bin_cropped_80 <- mask(Select_TSS_stack_bin_cropped_80, Andes_East.rasterized)
    Select_Jaccard_stack_bin_cropped_95 <- mask(Select_Jaccard_stack_bin_cropped_95, Andes_East.rasterized)
    Select_TSS_stack_bin_cropped_95 <- mask(Select_TSS_stack_bin_cropped_95, Andes_East.rasterized)
    
  } else {
    if (Andes_side == "West") {
      
      Select_Jaccard_stack_cont_cropped_80 <- mask(Select_Jaccard_stack_cont_cropped_80, Andes_West.rasterized)
      Select_TSS_stack_cont_cropped_80 <- mask(Select_TSS_stack_cont_cropped_80, Andes_West.rasterized)
      Select_Jaccard_stack_cont_cropped_95 <- mask(Select_Jaccard_stack_cont_cropped_95, Andes_West.rasterized)
      Select_TSS_stack_cont_cropped_95 <- mask(Select_TSS_stack_cont_cropped_95, Andes_West.rasterized)
      
      Select_Jaccard_stack_bin_cropped_80 <- mask(Select_Jaccard_stack_bin_cropped_80, Andes_West.rasterized)
      Select_TSS_stack_bin_cropped_80 <- mask(Select_TSS_stack_bin_cropped_80, Andes_West.rasterized)
      Select_Jaccard_stack_bin_cropped_95 <- mask(Select_Jaccard_stack_bin_cropped_95, Andes_West.rasterized)
      Select_TSS_stack_bin_cropped_95 <- mask(Select_TSS_stack_bin_cropped_95, Andes_West.rasterized)
    }
  }
  
  ### Add 0 value for terrestrial pixels to be able to distinguish from oceans
  Select_Jaccard_stack_cont_cropped_80 <- merge(x = continent_mask, y = Select_Jaccard_stack_cont_cropped_80)
  Select_Jaccard_stack_cont_cropped_95 <- merge(x = continent_mask, y = Select_Jaccard_stack_cont_cropped_95)
  Select_TSS_stack_cont_cropped_80 <- merge(x = continent_mask, y = Select_TSS_stack_cont_cropped_80)
  Select_TSS_stack_cont_cropped_95 <- merge(x = continent_mask, y = Select_TSS_stack_cont_cropped_95)
  
  Select_Jaccard_stack_bin_cropped_80 <- merge(x = continent_mask, y = Select_Jaccard_stack_bin_cropped_80)
  Select_Jaccard_stack_bin_cropped_95 <- merge(x = continent_mask, y = Select_Jaccard_stack_bin_cropped_95)
  Select_TSS_stack_bin_cropped_80 <- merge(x = continent_mask, y = Select_TSS_stack_bin_cropped_80)
  Select_TSS_stack_bin_cropped_95 <- merge(x = continent_mask, y = Select_TSS_stack_bin_cropped_95)
  
  
  # Save stacks as .RData/rds
  save(Select_Jaccard_stack_cont_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont_cropped_80.RData"), version = "2")
  saveRDS(Select_Jaccard_stack_cont_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont_cropped_80.rds"), version = "2")
  save(Select_TSS_stack_cont_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont_cropped_80.RData"), version = "2")
  saveRDS(Select_TSS_stack_cont_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont_cropped_80.rds"), version = "2")
  
  save(Select_Jaccard_stack_cont_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont_cropped_95.RData"), version = "2")
  saveRDS(Select_Jaccard_stack_cont_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_cont_cropped_95.rds"), version = "2")
  save(Select_TSS_stack_cont_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont_cropped_95.RData"), version = "2")
  saveRDS(Select_TSS_stack_cont_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_cont_cropped_95.rds"), version = "2")
  
  save(Select_Jaccard_stack_bin_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin_cropped_80.RData"), version = "2")
  saveRDS(Select_Jaccard_stack_bin_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin_cropped_80.rds"), version = "2")
  save(Select_TSS_stack_bin_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin_cropped_80.RData"), version = "2")
  saveRDS(Select_TSS_stack_bin_cropped_80, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin_cropped_80.rds"), version = "2")
  
  save(Select_Jaccard_stack_bin_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin_cropped_95.RData"), version = "2")
  saveRDS(Select_Jaccard_stack_bin_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_Jaccard_stack_bin_cropped_95.rds"), version = "2")
  save(Select_TSS_stack_bin_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin_cropped_95.RData"), version = "2")
  saveRDS(Select_TSS_stack_bin_cropped_95, file = paste0("../outputs/By_unit/",unit,"/Select_TSS_stack_bin_cropped_95.rds"), version = "2")
  
  
  ### 4/ Plot the final cropped EM ####

  # Function adding the points to the plot on each layer
  add_points <- function() {
    plot(unit_points2, add = TRUE, col = "red", pch = 16, cex = 0.3)
  }

  # 4.1/ Plot the 4 EM for 1 Eval metric and 1 Buffer, on a single pdf page
  all_EM_list <- list(all_EM_Jaccard_buffer_80, all_EM_Jaccard_buffer_95, all_EM_TSS_buffer_80, all_EM_TSS_buffer_95)
  all_EM_names_list <- c("all_EM_Jaccard_buffer_80", "all_EM_Jaccard_buffer_95", "all_EM_TSS_buffer_80", "all_EM_TSS_buffer_95")

  for (j in 1:length(all_EM_list))
  {
    EM_stack <- all_EM_list[[j]]
    names(EM_stack) <- c("Continuous", "Incertitude (IQR)", "Binary", "Committee Averaging")
    EM_stack_name <- all_EM_names_list[j]
    pdf(file = paste0("../Maps/By_unit/",unit,"/",EM_stack_name,"_",unit,".pdf"), height = 6, width = 7)
    internal_margins <- par()$mar
    external_margins <- par()$oma
    par(mar = c(5.1,6.1,9.1,3.1))
    par(oma = c(1,2,4,3))
    plot(EM_stack, addfun = add_points)
    title(main = unit, outer = T)
    par(mar = internal_margins, oma = external_margins)
    dev.off()
    # Copy in species folder
    # sp <- as.character(modeled_unit$Sp_full[i])
    file.copy(from = paste0("../Maps/By_unit/",unit,"/",EM_stack_name,"_",unit,".pdf"), to = paste0("../Maps/By_species/",sp,"/",EM_stack_name,"_",unit,".pdf"), overwrite = T)
  }

  # 4.2/ Plot the full projection and the two cropped one, for continuous and binary maps

  # Jaccard
  Final_stack <- stack(Ensemble_Jaccard_median,   # Full continuous map
                       Ensemble_Jaccard_bin,      # Full binary map
                       all_EM_Jaccard_buffer_80[["Ensemble_Jaccard_median_cropped_80"]],  # Buffer 80 continuous map
                       all_EM_Jaccard_buffer_80[["Ensemble_Jaccard_bin_cropped_80"]],     # Buffer 80 binary map
                       all_EM_Jaccard_buffer_95[["Ensemble_Jaccard_median_cropped_95"]],  # Buffer 95 continuous map
                       all_EM_Jaccard_buffer_95[["Ensemble_Jaccard_bin_cropped_95"]])     # Buffer 95 binary map

  names(Final_stack) <- c("Continuous Map", "Binary Map", "Continuous Map\nBuffer 80%", "Binary Map\nBuffer 80%", "Continuous Map\nBuffer 95%", "Binary Map\nBuffer 95%")

  pdf(file = paste0("../Maps/By_unit/",unit,"/All_buffers_cont_bin_Jaccard_",unit,".pdf"), height = 9, width = 7)
  internal_margins <- par()$mar
  external_margins <- par()$oma
  par(mar = c(5.1,6.1,9.1,3.1))
  par(oma = c(1,2,4,3))
  plot(Final_stack, addfun = add_points, nc = 2)
  title(main = unit, outer = T)
  par(mar = internal_margins, oma = external_margins)
  dev.off()
  # Copy in species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/All_buffers_cont_bin_Jaccard_",unit,".pdf"), to = paste0("../Maps/By_species/",sp,"/All_buffers_cont_bin_Jaccard_",unit,".pdf"), overwrite = T)

  # TSS
  Final_stack <- stack(Ensemble_TSS_median,   # Full continuous map
                       Ensemble_TSS_bin,      # Full binary map
                       all_EM_TSS_buffer_80[["Ensemble_TSS_median_cropped_80"]],  # Buffer 80 continuous map
                       all_EM_TSS_buffer_80[["Ensemble_TSS_bin_cropped_80"]],     # Buffer 80 binary map
                       all_EM_TSS_buffer_95[["Ensemble_TSS_median_cropped_95"]],  # Buffer 95 continuous map
                       all_EM_TSS_buffer_95[["Ensemble_TSS_bin_cropped_95"]])     # Buffer 95 binary map

  names(Final_stack) <- c("Continuous Map", "Binary Map", "Continuous Map\nBuffer 80%", "Binary Map\nBuffer 80%", "Continuous Map\nBuffer 95%", "Binary Map\nBuffer 95%")

  pdf(file = paste0("../Maps/By_unit/",unit,"/All_buffers_cont_bin_TSS_",unit,".pdf"), height = 9, width = 7)
  internal_margins <- par()$mar
  external_margins <- par()$oma
  par(mar = c(5.1,6.1,9.1,3.1))
  par(oma = c(1,2,4,3))
  plot(Final_stack, addfun = add_points, nc = 2)
  title(main = unit, outer = T)
  par(mar = internal_margins, oma = external_margins)
  dev.off()
  # Copy in species folder
  # sp <- as.character(modeled_unit$Sp_full[i])
  file.copy(from = paste0("../Maps/By_unit/",unit,"/All_buffers_cont_bin_TSS_",unit,".pdf"), to = paste0("../Maps/By_species/",sp,"/All_buffers_cont_bin_TSS_",unit,".pdf"), overwrite = T)

  cat(paste0("\n", Sys.time()," ------ Done for ", unit, " = Unit N°",i,"/",nrow(modeled_unit)," ------\n"))
 
} 

# Save Summary table for unit models
save(list_occ, file = paste0(".././input_data/occurences_df/list_occ_after_clipping.RData"), version = "2")
saveRDS(list_occ, file = paste0("../input_data/occurences_df/list_occ_after_clipping.rds"), version = "2")


# plot(list_occ$Uncertainty.80 ~ list_occ$Uncertainty.95)
# 
# hist(list_occ$Uncertainty.80)
# hist(list_occ$Uncertainty.95)

```

```{r}
pdf(file="../outputs/full_pdf_after_clipping.pdf", height = 9, width = 7)

p<-progress_estimated(nrow(modeled_unit), min_time=10)
  
for(i in 1:nrow(modeled_unit))
{
  p$pause(0.1)$tick()$print()
  unit<-modeled_unit$biomodTag[i]
  unit2<-modeled_unit$code_unit[i]
  EP80<-readRDS(file = paste0("../outputs/By_unit/",unit,"/Ensemble_Jaccard_median_cropped_80.rds"))
  load(paste0("../input_data/Species_data/Spatial_Points_Objects/occurrences_",unit2,".RData"))
  
  plot(EP80, main=unit)
  points(x= unit_points2$x, y= unit_points2$y, pch = 16, cex = 0.3, col = "red")
}
dev.off
```
