==> scripts/19b_Index_after_review.Rmd <==
---
title: "19b_Index_after_review"
author: "Eddie Pérochon"
date: "2025-05-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

##### Script 19b : Compute new indices maps  #####

###################################################
#      Authors: Eddie Pérochon & Maël Doré        #
#      Contact: eddie.perochon@hotmail.com        #
###################################################


### Goals = 
## Compute biodiversity index maps recommended by reviewers

# 
1/ Heliconiini Phylogenetic diversity LOESS regression
2/ Ithomiini Phylogenetic diversity LOESS regression

### Inputs 
    # Stack of species richness and PD from script 19 (Heliconiini)
    # Stack of species richness and PD from Doré et al. 2022

### Outputs
    # LOESS regression outputs
    # New divrersity maps for main text and supplementary
    
```{r}
### Prepare stuff

# Effacer l'environnement
rm(list = ls())

library(raster)
library(rangeBuilder)
library(gplots)
library(Rarity)
library(ape)
library(picante)
library(geiger)
library(treespace)
library(rgeos)
library(sf)
library(dplyr)
library(stringr)
library(biscale)
library(cowplot)
library(sf)
library(ggplot2)
library(bivariatemaps)
library(png)
library(phytools)
library(TreeTools)
```

```{r}
#Load index
h_ind <- readRDS( "../outputs/Indices_stacks/All_indices_Heliconiini.RDS")

pal_bl_red_Mannion <- readRDS(file = "../Maps/pal_bl_red_Mannion.rds")

# Load mask for continent borders, plot border, and grid
grid_Mollweide_out <- readRDS(file = "../input_data/Biogeographic_data/grid_Mollweide_out.rds")
large_bg_mask_Mollweide <- readRDS(file = "../input_data/envData/crop_shape.RDS")
bbox_sp_Mollweide <- readRDS(file = "../input_data/Biogeographic_data/bbox_sp_Mollweide.rds")

# Load stuff for plot on Andes
Andes_ext <- extent(c(-90, -59, -15, 14))
Close_Andes_ext <- extent(c(-90, -67.2, -18, 14))

envData <- readRDS(file = paste0("../input_data/envData/Var_stack_WGS84.RDS"))
DEM <- envData[[5]]
Andes_DEM <- crop(DEM, Andes_ext)
Close_Andes_DEM <- crop(DEM, Close_Andes_ext)
Andes_bbox <- as(extent(Andes_DEM), 'SpatialPolygons')
Andes_bbox@proj4string@projargs <- "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs" 

Andes_bbox<-spTransform(Andes_bbox, CRSobj = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
Andes_DEM<-projectRaster(Andes_DEM, crs=CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
Close_Andes_DEM<-projectRaster(Close_Andes_DEM, crs=CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))

saveRDS(Andes_bbox, file = "../input_data/Biogeographic_data/Andes_bbox.rds", version = "2")


DEM<-readRDS("../input_data/envData/Envstack.rds")[[5]] #goback to moll

sea_shp<-readRDS(file="../input_data/Biogeographic_data/sea_shp.RDS")

### Load maps 
sp_richness <- readAll(readRDS(file = paste0("../outputs/Indices_maps/tot.sp.richness_Jaccard.80.rds")))
# sp_mean_geo_rarity <- readRDS(file = "../outputs/Indices_maps/sp.mean.rarity_Jaccard.80.rds")       # Linear weighting
sp_mean_geo_rarity <- readAll(readRDS(file = "../outputs/Indices_maps/sp.mean.rarity_Leroy_Jaccard.80_contrasted.rds"))   # Leroy's weighting
Faith_PD <- readAll(readRDS(file = "../outputs/Indices_Maps/PD.raster_Jaccard.80.rds"))

load(file = paste0("../outputs/Indices_maps/ring.richness_Jaccard_ss.80.Rdata"))
ring_richness_ss <-readAll(ring.richness_Jaccard.80)
load(file = paste0("../outputs/Indices_maps/ring.richness_Jaccard_sl.80.Rdata"))
ring_richness_sl <-readAll(ring.richness_Jaccard.80)


# ring_mean_geo_rarity <- readRDS(file = "../outputs/Indices_maps/mimicry.mean.rarity_Jaccard.80.rds")  # Linear weighting
ring_mean_geo_rarity_ss <- readAll(readRDS(file = "../outputs/Indices_maps/ring.mean.rarity_Leroy_Jaccard_ss.80_contrasted.rds")) 
ring_mean_geo_rarity_sl <- readAll(readRDS(file = "../outputs/Indices_maps/ring.mean.rarity_Leroy_Jaccard_sl.80_contrasted.rds"))
# Leroy's weighting
weighted_mean_ring_size_ss <- readAll(readRDS(file = "../outputs/Indices_maps/weighted_mean_ring_size_ss.rds"))
weighted_mean_ring_size_sl <- readAll(readRDS(file = "../outputs/Indices_maps/weighted_mean_ring_size_sl.rds"))

```

```{r}
#Heliconiini LOESS regression
h_sprv <- values(sp_richness)
h_pdv <- values(Faith_PD)

hloess <- loess(h_pdv~h_sprv)
hloess_v <- h_pdv
hloess_v[!is.na(hloess_v)] <- residuals(hloess)

#Standardize residuals
# hloess_v <- scale(hloess_v)
hloess_v[h_sprv == 0] <- NA

hloess_map <- sp_richness
values(hloess_map) <- hloess_v

plot(hloess_map)

#Save index
saveRDS(hloess_map ,file = "../outputs/Indices_maps/heliconiini_pd_loess.rds")

```

```{r}
### 2/ Generate Andes maps ####

# 2.1/ Function to crop and save maps

crop_to_Andes <- function(x) # Raster to project
{
  x_name <- deparse(substitute(x)) # Get the name of the initial raster as a character string
  
  # Crop into Andes
  new_map <- crop(x, Andes_DEM)
  
    # Generate new object with "Andes_" prefix in the global environment
  eval(call("<<-", as.name(paste0("Andes_", x_name)), new_map))
  
  # Save new object in outputs folder
  saveRDS(new_map, file = paste0("../outputs/Indices_maps/Andes/Andes_", x_name, ".rds"))
}


```


```{r}
# 2.2/ Crop maps

crop_to_Andes(sp_richness)
crop_to_Andes(ring_mean_geo_rarity_ss)
crop_to_Andes(ring_mean_geo_rarity_sl)
crop_to_Andes(weighted_mean_ring_size_ss)
crop_to_Andes(weighted_mean_ring_size_sl)
crop_to_Andes(sea_shp)
crop_to_Andes(grid_Mollweide_out)
```


```{r}
# 3.2/ Project all maps

# Put all maps in a list

list_all_maps <- list(sp_richness, sp_mean_geo_rarity, Faith_PD, hloess_map, ring_richness_ss, ring_mean_geo_rarity_ss, weighted_mean_ring_size_ss, Andes_sp_richness, Andes_ring_mean_geo_rarity_ss, Andes_weighted_mean_ring_size_ss, ring_richness_sl, ring_mean_geo_rarity_sl, weighted_mean_ring_size_sl,  Andes_ring_mean_geo_rarity_sl, Andes_weighted_mean_ring_size_sl)
names(list_all_maps) <- c("sp_richness", "sp_mean_geo_rarity", "Faith_PD", "hloess","ring_richness_ss", "ring_mean_geo_rarity_ss", "weighted_mean_ring_size_ss"," Andes_sp_richness", "Andes_ring_mean_geo_rarity_ss", "Andes_weighted_mean_ring_size_ss", "ring_richness_sl", "ring_mean_geo_rarity_sl", "weighted_mean_ring_size_sl",  "Andes_ring_mean_geo_rarity_sl", "Andes_weighted_mean_ring_size_sl")


```

```{r}
### 4/ Plot final figure ####

# 4.1/ Plotting function ####

{
  map_indices_Mollweide <- function(x,                                    # Raster to map
                                  color_palette = pal_bl_red_Mannion,   # Color palette
                                  main_title,                           # Main title
                                  main_title_cex = 1.5,                 # Main title size
                                  
                                  xlim = c(-5457, 4953),   # Limit of plot on x-axis (Longitude)
                                  ylim = c(-4495, 5705),    # Limit of plot on y-axis (Latitude)
                                  axis_cex = 1.4,             # Axes size
                                  
                                  xlab = "",                # X-axis label
                                  ylab = "",                # Y-axis label
                                  x_axis_breaks = c(-3930, -2170, -420, 1500, 3050),            # X-axis tick breaks
                                  y_axis_breaks = c(-3650, -2450, -1220, 0, 1230, 2445, 3660, 4890),        # Y-axis tick breaks
                                  x_axis_labels = c("120°E", "100°E", "80°E", "60°E", "40°E"),      # X-axis tick labels
                                  y_axis_labels = c("30°S", "20°S", "10°S", "0°", "10°N", "20°N", "30°N", "40°N"),  # Y-axis tick labels
                                  
                                  legend_title,             # Legend title
                                  legend_title_cex = 1.4,   # Legend title size
                                  legend_title_x = -3550,   # Legend title x position
                                  legend_title_y = 430,     # Legend title y position
                                  legend_cex = 1.4,         # Legend size
                                  legend_location = c(-4100, -3800, -3950, 0),  # Legend position
                                  
                                  scale_bar_position = c(-2600, -4000),  # Scale bar position
                                  
                                  arrow_scale = 0.45,           # North arrow size
                                  arrow_padin = c(0.15, 0.15),  # North arrow position adjustement
                                  
                                  facet_letter = "",                  # Small case letter for facet
                                  facet_letter_col = "black",         # Color of case letter for facet
                                  facet_letter_cex = 2.2,             # Size of small case letter for facet
                                  facet_letter_inset = c(0, -0.008),  # Position adjustment of small case letter for facet,
  legend_breaks = seq(cellStats(x, stat = "min"), cellStats(x, stat = "max"), length.out = 5)) 

{
  
  legend_breaks <- round(legend_breaks, 1)
    
  # Plot raster background without axis
  image(sp_richness, col = "#EDEDED", axes = F)
  par(new = TRUE)
  image(x, col = color_palette,
        xlim = xlim, ylim = ylim, axes = F,
        xlab = xlab, ylab = ylab)
  title(main = main_title, cex.main = main_title_cex, line = 1)
  
  # Generate axes with manual positioning of ticks
  axis(side = 1, at = x_axis_breaks, labels = x_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0, padj = 0.5)
  axis(side = 2, at = y_axis_breaks, labels = y_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0)
  
  # Add background, borders and graticules
  plot(sea_shp, legend=F, col="#E1f1f1", add=T )
  plot(large_bg_mask_Mollweide, lwd = 1, col = "grey20", add = T)
  plot(grid_Mollweide_out, lty = 92, col = "grey80",legend=F,  add = T)
  plot(bbox_sp_Mollweide, lwd = 2, border = "black", col = NA, legend=F, add = T)

  
  # Add scale bar in legend
  scalebar(d = 2000, type = "line", lwd = 4, divs = 4, xy = scale_bar_position, label = c("", "2000 km", ""), adj = c(0.5, -0.8), font = 2, cex = 1.2)
  prettymapr::addnortharrow(scale = arrow_scale, padin = arrow_padin, text.col = "#00000000")
  epm::addLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
  epm::addLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
  graphics::text(x = legend_title_x, y = legend_title_y, font = 2, cex = legend_title_cex, label = legend_title)
  
  # Add facet letter
  legend(legend = facet_letter, x = "bottomright", bty = "n", text.col = facet_letter_col,
         text.font = 2, cex = facet_letter_cex, inset = facet_letter_inset)
  
  }
}

{
  map_indices_Andes_Mollweide <- function(x,                                    # Raster to map
                                          color_palette = pal_bl_red_Mannion,   # Color palette
                                          main_title,                           # Main title
                                          main_title_cex = 1.5,                 # Main title size
                                          
                                          xlim = c(-1654, 1760),    # Limit of plot on x-axis (Longitude)
                                          ylim = c(-2014, 1879),    # Limit of plot on y-axis (Latitude)
                                          axis_cex = 1.4,             # Axes size
                                          
                                          xlab = "",                # X-axis label
                                          ylab = "",                # Y-axis label
                                          x_axis_breaks = c(-1550, -420, 520, 1500),
                                          x_axis_labels = c("90°E", "80°E", "70°E", "60°E"),
                                          y_axis_breaks = c(-1850, -1220, -600, 0, 610, 1230), 
                                          y_axis_labels = c("15°S", "10°S", "5°S", "0°", "5°N", "10°N"),
                                          
                                          legend_title,             # Legend title
                                          legend_title_cex = 1.4,   # Legend title size
                                          legend_title_x = -1300,   # Legend title x position
                                          legend_title_y = 840,     # Legend title y position
                                          legend_cex = 1.4,            # Legend size
                                          legend_location = c(-1460, -1350, -1250, 640),  # Legend position
                                          
                                          elevation = TRUE, # Should elevation contour and assoicated legend added?
                                          
                                          scale_bar_position = c(-800, -1900),  # Scale bar position
                                          
                                          arrow_scale = 0.45,           # North arrow size
                                          arrow_padin = c(0.15, 0.15),  # North arrow position adjustement
                                          
                                          facet_letter = "",                  # Small case letter for facet
                                          facet_letter_col = "black",         # Color of case letter for facet
                                          facet_letter_cex = 2.2,             # Size of small case letter for facet
                                          facet_letter_inset = c(0, -0.008))  # Position adjustment of small case letter for facet
  
  {
      legend_breaks <- seq(cellStats(x, stat = "min"), cellStats(x, stat = "max"), length.out = 5)  
      legend_breaks <- round(legend_breaks, 1)
    # Plot raster background without axis
    image(x, col = color_palette,
          xlim = xlim, ylim = ylim, axes = F,
          xlab = xlab, ylab = ylab)
    title(main = main_title, cex.main = main_title_cex, line = 1)
    
    # Generate axes with manual positioning of ticks
    axis(side = 1, at = x_axis_breaks, labels = x_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0, padj = 0.5)
    axis(side = 2, at = y_axis_breaks, labels = y_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0)
    
    # Add background, borders and graticules
    plot(Andes_sea_shp, legend=F, col="#E1f1f1", add=T)
    plot(large_bg_mask_Mollweide, lwd = 1, col = "grey20", add = T)
    plot(Andes_grid_Mollweide_out, lty = 92, col = "grey80",legend=F, add = T)
    
    # plot(Andes_bbox, lwd = 2, border = "black", col = NA, add = T)
   
    if (elevation == TRUE) 
    {
      # Add elevation contour 
      contour(x = Close_Andes_DEM, levels = c(500, 2500), lwd = 1.5, lty = c(2, 1),
              drawlabels = F, col = "black", add = T)
      legend(x = "bottomleft", title = "", legend = c("2500m", "500m"), lty = c(1, 2) , lwd = 2, bty = "n", cex = 1.0, text.font = 2)
      graphics::text(x = -1300, y = -1500, font = 2, cex = 1.1, label = "Elevation")
    }
    
    # Add scale bar in legend
    scalebar(d = 500, type = "line", lwd = 3.5, divs = 4, xy = scale_bar_position, label = c("", "500 km", ""), adj = c(0.5, -0.8), font = 2, cex = 1.2)
    prettymapr::addnortharrow(scale = arrow_scale, padin = arrow_padin, text.col = "#00000000")
    epm::addLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
    epm::addLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
    graphics::text(x = legend_title_x, y = legend_title_y, font = 2, cex = legend_title_cex, label = legend_title)
    
    # Add facet letter
    legend(legend = facet_letter, x = "bottomright", bty = "n", text.col = facet_letter_col,
           text.font = 2, cex = facet_letter_cex, inset = facet_letter_inset)
    
  }
}

  map_indices_Andes_Mollweide_down <- function(x,                                    # Raster to map
                                    color_palette = pal_bl_red_Mannion,   # Color palette
                                    main_title,                           # Main title
                                    main_title_cex = 1.5,                 # Main title size
                                    
                                    xlim = c(-1654, 1760),    # Limit of plot on x-axis (Longitude)
                                    ylim = c(-2014, 1879),    # Limit of plot on y-axis (Latitude)
                                    axis_cex = 1.4,             # Axes size
                                    
                                    xlab = "",                # X-axis label
                                    ylab = "",                # Y-axis label
                                    x_axis_breaks = c(-1550, -420, 520, 1500),
                                    x_axis_labels = c("90°E", "80°E", "70°E", "60°E"),
                                    y_axis_breaks = c(-1850, -1220, -600, 0, 610, 1230), 
                                    y_axis_labels = c("15°S", "10°S", "5°S", "0°", "5°N", "10°N"),
                                    
                                    legend_title,             # Legend title
                                    legend_title_cex = 1.4,   # Legend title size
                                    legend_title_x = -1300,   # Legend title x position
                                    legend_title_y = 290,     # Legend title y position
                                    legend_cex = 1.4,         # Legend size
                                            # Legend tick positions
                                    legend_location = c(-1460, -1350, -1800, 90),  # Legend position
                                    
                                    elevation = TRUE,            # Should elevation contour and associated legend added?
                                    y_adjust_lvl = 0,            # Adjustment for elevation level legend
                                    y_adjust_elevation = 0,      # Adjustment for elevation label
                                    elevation_title_cex = 1.3,   # Size of Elevation title
                                    elevation_legend_cex = 1.2,  # Size of Elevation legend
                                    
                                    scale_bar_position = c(-1000, -1800),  # Scale bar position
                                    
                                    arrow_scale = 0.45,           # North arrow size
                                    arrow_padin = c(0.15, 0.15),  # North arrow position adjustment
                                    
                                    facet_letter = "",                  # Small case letter for facet
                                    facet_letter_col = "black",         # Color of case letter for facet
                                    facet_letter_cex = 2.2,             # Size of small case letter for facet
                                    facet_letter_inset = c(0, -0.008))  # Position adjustment of small case letter for facet
  
  {
          legend_breaks <- seq(cellStats(x, stat = "min"), cellStats(x, stat = "max"), length.out = 5)  
      legend_breaks <- round(legend_breaks, 1)
    # Plot raster background without axis
    image(x, col = color_palette,
          xlim = xlim, ylim = ylim, axes = F,
          xlab = xlab, ylab = ylab)
    title(main = main_title, cex.main = main_title_cex, line = 1)
    
    # Generate axes with manual positioning of ticks
    axis(side = 1, at = x_axis_breaks, labels = x_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0, padj = 0.5)
    axis(side = 2, at = y_axis_breaks, labels = y_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0)
    
    # Add background, borders and graticules
    plot(Andes_sea_shp, legend=F, col="#E1f1f1", add=T)
    plot(large_bg_mask_Mollweide, lwd = 1, col = "grey20", add = T)
    plot(Andes_grid_Mollweide_out, lty = 92, col = "grey80", legend=F, add = T)
    #plot(Andes_bbox, lwd = 2, border = "black", col = NA, add = T)
    
    
    if (elevation == TRUE) 
    {
      # Add elevation contour 
      contour(x = Close_Andes_DEM, levels = c(500, 2500), lwd = 1.5, lty = c(2, 1),
              drawlabels = F, col = "black", add = T)
      # Add legend for elevation
      legend(x = "bottomleft", inset = c(0, 0.63 + y_adjust_lvl), title = "", legend = c("2500m", "500m"), lty = c(1, 2) , lwd = 2, bty = "n", cex = elevation_legend_cex, text.font = 2)
      graphics::text(x = -1250, y = 1040 + y_adjust_elevation, font = 2, cex = elevation_title_cex, label = "Elevation")
    }

    # Add scale bar in legend
    scalebar(d = 500, type = "line", lwd = 3.5, divs = 4, xy = scale_bar_position, label = c("", "500 km", ""), adj = c(0.5, -0.8), font = 2, cex = 1.2)
    prettymapr::addnortharrow(scale = arrow_scale, padin = arrow_padin, text.col = "#00000000")
    epm::addLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
    epm::addLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
    graphics::text(x = legend_title_x, y = legend_title_y, font = 2, cex = legend_title_cex, label = legend_title)
    
    # Add facet letter
    legend(legend = facet_letter, x = "bottomright", bty = "n", text.col = facet_letter_col,
           text.font = 2, cex = facet_letter_cex, inset = facet_letter_inset)
    
  }

```


```{r}
for(m in c("ss", "sl"))
{
pdf(file = paste0("../Maps/Index_maps_Mollweide_aftRev_",m,".pdf"), height = 12, width = 12)

internal_margins <- par()$mar
# par(mar = c(3.1, 3.5, 3.5, 2.1))
par(mar = c(3.1, 3.1, 2.7, 1.6))
par(mfrow = c(3, 3))

# A/ Species richness ####

map_indices_Mollweide(x = sp_richness,
                      main_title = "Species richness",
                      legend_title = "Species",
                      legend_title_x = -3650,
                      legend_title_y = 430,
                      # legend_breaks = seq(0, 35, 5), 
                      facet_letter = "(a)")

# B/ Species mean geographic rarity ####

# Need to add a few values to lower values to avoid being colored as null outer-range values
sp_mean_geo_rarity[(sp_mean_geo_rarity[] > 0) & (sp_mean_geo_rarity[] <= 0.005)] <- 0.005

map_indices_Mollweide(x = sp_mean_geo_rarity,
                      main_title = "Species geographic rarity",
                      legend_title = "Rarity\nindex",
                      legend_title_x = -3650,
                      legend_title_y = 870,
                      # legend_breaks = seq(0, 0.5, 0.1), 
                      facet_letter = "(b)")

# C/ Faith's PD ####

map_indices_Mollweide(x = hloess_map,
                      main_title = "Phylogenetic diversity ",
                      legend_title = "Residuals",
                      legend_title_x = -3650,
                      legend_title_y = 670,
                      # legend_breaks = seq(1, 290, 70), 
                      facet_letter = "(c)")

# D/ Mimicry richness ####
ring_richness <- get(paste0("ring_richness_",m))

map_indices_Mollweide(x = ring_richness,
                      main_title = "Mimicry richness",
                      legend_title = "Mimicry\nrings",
                      legend_title_x = -3670,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 24, 4), 
                      facet_letter = "(d)")

# E/ Mimicry ring mean geographic rarity ####
ring_mean_geo_rarity <- get(paste0("ring_mean_geo_rarity_",m))

map_indices_Mollweide(x = ring_mean_geo_rarity,
                      main_title = "Mimicry geographic rarity",
                      legend_title = "Rarity\nindex",
                      legend_title_x = -3650,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 0.44, 0.1), 
                      facet_letter = "(e)")

# F/ Mean ring size ####
weighted_mean_ring_size <- get(paste0("weighted_mean_ring_size_",m))

map_indices_Mollweide(x = weighted_mean_ring_size,
                      main_title = "Mean mimicry ring size",
                      legend_title = "Ring\nsize",
                      legend_title_x = -3750,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 3, 0.5), 
                      facet_letter = "(f)")

# G/ Species richness in Andes ####

map_indices_Andes_Mollweide_down(x = sp_richness,
                            main_title = "Species richness",
                            legend_title = "Species",
                            y_adjust_lvl = 0.0,            # Adjustment for elevation level legend
                            y_adjust_elevation = 0, 
                            # legend_breaks = seq(0, 35, 5), 
                            facet_letter = "(g)",
                            facet_letter_col = "white")

# H/ Mean species geographic rarity in Andes ####

map_indices_Andes_Mollweide_down(x = sp_mean_geo_rarity,
                                 color_palette = pal_bl_red_Mannion[-1],
                                 main_title = "Species geographic rarity",
                                 legend_title = "Rarity index",
                                 legend_title_x = -1100,
                                 legend_title_y = 350,
                                 y_adjust_lvl = 0.0,
                                 y_adjust_elevation = 0,
                                 # legend_breaks = seq(0, 0.5, 0.1), 
                                 facet_letter = "(h)",
                                 facet_letter_col = "white")

# I/ Mean ring size in Andes ####
weighted_mean_ring_size <- get(paste0("weighted_mean_ring_size_",m))

# map_indices_Andes_Mollweide_down(x = weighted_mean_ring_size_Mollweide,
#                             main_title = "Mean mimicry ring size",
#                             legend_title = "Ring\nsize",
#                             legend_title_x = -1330,
#                             legend_title_y = 400,
#                             elevation = FALSE,
#                             legend_breaks = seq(0, 6, 1), 
#                             facet_letter = "(i)",
#                             facet_letter_col = "white")
# 
# plot(Central_America_shp2_Mollweide, col = NA, border = "#F8766D", lwd = 4, add = T)
# plot(East_Ecuador_shp2_Mollweide, col = NA, border = "#7CAE00", lwd = 4, add = T)
# plot(Peruvian_shp2_Mollweide, col = NA, border = "#00BFC4", lwd = 4, add = T)
# 
# graphics::text(x = -570, y = 1500, font = 2, cex = 1.4, label = "CA", col = "#F8766D")
# graphics::text(x = -780, y = 10, font = 2, cex = 1.4, label = "EE", col = "#7CAE00")
# graphics::text(x = -640, y = -1350, font = 2, cex = 1.4, label = "EP", col = "#00BFC4")
# 
# segments(x0 = -610, y0 = 1350 , x1 = -700, y1 = 1100, col = "#F8766D", lwd = 4)
# segments(x0 = -620, y0 = -20 , x1 = -400, y1 = -100, col = "#7CAE00", lwd = 4)
# segments(x0 = -490, y0 = -1320 , x1 = -250, y1 = -1150, col = "#00BFC4", lwd = 4)

map_indices_Andes_Mollweide_down(x = weighted_mean_ring_size,
                                 main_title = "Mean mimicry ring size",
                                 legend_title = "Ring size",
                                 legend_title_x = -1200,
                                 # legend_title_y = 350,
                                 elevation = TRUE,
                                 # legend_breaks = seq(0, 3, 0.5), 
                                 facet_letter = "(i)",
                                 facet_letter_col = "white")

par(mar = internal_margins)
par(mfrow = c(1, 1))

dev.off()
}

```

```{r}
#Ithomiini LOESS regression
### Load maps 

for(m in c("ss", "sl"))
{
####Helico
sp_richness_H <- readRDS(file = paste0("../outputs/Indices_maps/tot.sp.richness_Jaccard.80.rds"))
# sp_mean_geo_rarity <- readRDS(file = "../outputs/Indices_maps/sp.mean.rarity_Jaccard.80.rds")       # Linear weighting
sp_mean_geo_rarity_H <- readRDS(file = "../outputs/Indices_maps/sp.mean.rarity_Leroy_Jaccard.80_contrasted.rds")   # Leroy's weighting
Faith_PD_H <- readRDS(file = "../outputs/Indices_Maps/PD.raster_Jaccard.80.rds")

load(file = paste0("../outputs/Indices_maps/ring.richness_Jaccard_",m,".80.Rdata"))
ring_richness_H<-readAll(readRDS(paste0("../outputs/Indices_maps/ring.richness_Jaccard_",m,".80.rds")))
# ring_mean_geo_rarity <- readRDS(file = "../outputs/Indices_maps/mimicry.mean.rarity_Jaccard.80.rds")  # Linear weighting
ring_mean_geo_rarity_H <- readRDS(file = paste0("../outputs/Indices_maps/ring.mean.rarity_Leroy_Jaccard_",m,".80_contrasted.rds") )   # Leroy's weighting
weighted_mean_ring_size_H <- readRDS(file = paste0("../outputs/Indices_maps/weighted_mean_ring_size_",m,".rds"))

Full_indices_H <- stack(sp_richness_H, sp_mean_geo_rarity_H, hloess_map, ring_richness_H, ring_mean_geo_rarity_H, weighted_mean_ring_size_H)
names(Full_indices_H)<-c("sp_richness", "sp_mean_geo_rarity", "loess", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

for(i in 2:nlayers(Full_indices_H))
{
  Full_indices_H[[i]]@data@values[Full_indices_H[[i]]@data@values == 0 & is.na(Full_indices_H[[1]]@data@values) == FALSE]
}

saveRDS(Full_indices_H, paste0("../outputs/Full_Helico_indice_stack_",m,"_loess.RDS"))

}

#Itho

Full_indices_I <- readRDS( file="../Ithomiini/Full_indices_I.RDS")
names(Full_indices_I)<-c("sp_richness", "sp_mean_geo_rarity", "FaithPD", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")


#LOESS
#Heliconiini LOESS regression
i_sprv <- values(Full_indices_I[["sp_richness"]])
i_pdv <- values(Full_indices_I[["FaithPD"]])

iloess <- loess(i_pdv~i_sprv)
iloess_v <- i_pdv
iloess_v[!is.na(iloess_v)] <- residuals(iloess)

#Standardize residuals
# iloess_v <- scale(iloess_v)
iloess_v[i_sprv == 0] <- NA

iloess_map <- sp_richness
values(iloess_map) <- iloess_v

#Save index
saveRDS(iloess_map ,file = "../outputs/Indices_maps/ithomiini_pd_loess.rds")

Full_indices_I[["FaithPD"]] <- iloess_map
names(Full_indices_I)<-c("sp_richness", "sp_mean_geo_rarity", "loess", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

Full_indices_I<-projectRaster(from=Full_indices_I, crs = crs(Full_indices_H), method="bilinear")
Full_indices_I<-resample(Full_indices_I, Full_indices_H[[1]])
saveRDS(Full_indices_I, file="../Ithomiini/Full_indices_I_wLOESS.RDS")


Full_indices_H <- readRDS(paste0("../outputs/Full_Helico_indice_stack_",m,"_loess.RDS"))

Full_indices_I <- readRDS(file="../Ithomiini/Full_indices_I_wLOESS.RDS")
Full_indices_I[["loess"]]@data@values[is.na(Full_indices_I[["loess"]]@data@values) & !is.na(Full_indices_H[["loess"]]@data@values)] <- 0


```


```{r}
#Mollweide map, create legend for each map
names_indices_clean_lg <- c("Species\n Richness", "Mean Species\n Geographic Rarity", "Phylogenetic\nDiversity", "Mimicry ring\n Richness", "Mimicry ring\n Geographic Rarity", " Mimicry ring\n Mean Size")
names_indices_clean <- c("Species Richness", "Mean Species Geographic Rarity", "Phylogenetic Diversity", "Mimicry ring richness", "Mimicry ring Geographic Rarity", "Mimicry ring Mean Size")
names_indices <- c("sp_richness", "sp_mean_geo_rarity", "loess", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

#extract values


  source("../functions/colmat_modified.R")
source("../functions/add_bivar_legend.R")
  source("../functions/biv_map_modified.R")

for(m in c("ss", "sl"))

{
  Full_indices_H <-  readRDS(paste0("../outputs/Full_Helico_indice_stack_",m,"_loess.RDS"))
  


#Number of quantiles
nquantiles = 4

for(i in 1:length(names_indices_clean))
  {
  names <- names_indices_clean[i]
  

  #Yellow Blue, colorblind friendly
  col.matrix <- colmat_modif(nquantiles=4, upperleft="#f3b300", upperright="#000000", bottomleft="#f3f3f3", bottomright="#509dc2", xlab=paste0("Heliconiini ", names), ylab= paste0("Ithomiini ", names), valuesx = Full_indices_H[[i]], valuesy =  Full_indices_I[[i]])

  bivmap <- biv_map_modified(Full_indices_H[[i]], Full_indices_I[[i]], col.matrix, nquantiles = 4)
  #   
  assign(paste0(names_indices[i], "_bivmap_",m), bivmap)
  
  #Add grey for 0x0 as last color
  col.matrix_v <- as.vector(col.matrix)
  col.matrix_v[ length(col.matrix) + 1] <-  "lightgrey" #"#DEE1EB"
  
  assign(paste0(names_indices[i], "_colmat"), col.matrix)
  assign(paste0(names_indices[i], "_colv"), col.matrix_v)
  
}

for(l in 1:6)
{
name_code <- names_indices[l]
col.matrix =   get(paste0(name_code, "_colmat"))

scale_decimals <- 1 #General case, not possible to have more numbers so for big number, no decimals

if(max(Full_indices_H[[l]]@data@values, na.rm= TRUE) > 9.99)
{
 scale_decimals <- 0 
}

add_bivar_legend(rasterx = Full_indices_H[[l]],
                 rastery = Full_indices_I[[l]],
                 color.matrix = col.matrix,
                 nquantiles = 4,
                 xlab = substitute(paste(bold("Heliconiini"))),
                 ylab = substitute(paste(bold("Ithomiini"))),
                 pngname = paste0("../Maps/Map_legends/",name_code,"_Bivmap_legend_",m,"_LOESS.png" ),
                 scaledeci = scale_decimals)
}                 
}

#Create lines for distribution borders

distrib_borders <- sp_richness_bivmap_ss
distrib_borders@data@values[distrib_borders@data@values > nquantiles^2] <- NA
distrib_borders@data@values[distrib_borders@data@values <= nquantiles^2] <- 1

distrib_borders <- boundaries(distrib_borders
                , type='outer')
distrib_borders@data@values[distrib_borders@data@values == 0] <- NA

distrib_borders <- mask(distrib_borders, sp_richness_bivmap_ss)

```


```{r}
#Mollweide map, final plot

source("../functions/custom_Bivariates_maps_mollweide.R")

for(m in c("ss", "sl"))
{

pdf(paste0("../Maps/BivarMap_",m,"_LOESS.pdf"), height = 8, width = 12)
par(mar = c(3.1, 3.1, 2.7, 1.6))
par(mfrow = c(2, 3))

for (l in 1:6)
{
name_code <- names_indices[l]
name_clean <- names_indices_clean[l]
name_clean_lg <- names_indices_clean_lg[l]

map <- get(paste0(name_code, "_bivmap_",m))
palette_bv <- unique(get(paste0(name_code, "_colv")))
  
 image(sea_shp,  col="#E1f1f1", #background
         xlim = c(-5457, 4953), ylim = c(-4495, 5705), axes = F,
         xlab = "", ylab = "")
# Add background, borders and graticules
  plot(bbox_sp_Mollweide, lwd = 2, border = "black", col = NA, legend=F, add = T)
  plot(grid_Mollweide_out, lty = 92, col = "grey80",legend=F,  add = T)
  par(new=TRUE)

Bivar_map_Mollweide(x = map, color_palette = palette_bv, main_title = name_clean, legend_title = name_clean_lg, facet_letter = paste0("(",letters[l],")") )
  plot(distrib_borders, col = "red", add = T, legend = F)
plot(large_bg_mask_Mollweide, lwd = 1, col = "grey20", add = T)
par(new=TRUE)

legend_biv <- readPNG(source = paste0("../Maps/Map_legends/",name_code,"_Bivmap_legend_",m,"_LOESS.png"))

par(new = TRUE)
rasterImage(legend_biv, xleft = -5500, ybottom = -4400, xright = -1000, ytop = 100)


}
dev.off()
}
```


```{r}
####Compute SESPD

#Heliconiini

com_mat_h <- readRDS(file = paste0("../outputs/Indices_stacks/sp.proba.mat_Jaccard.80.rds"))
phylo_helico <- read.nexus(file="../input_data/update_2024/Heliconiini_Chazot2019priors/Heliconiini_Chazot2019priors.MCC.median_ages.tree")

com_mat_h <- com_mat_h[,colnames(com_mat_h) %in% phylo_helico$tip.label]
row.names(com_mat_h) <- paste0("c",1:nrow(com_mat_h))
com_mat_h <- com_mat_h[rowSums(is.na(com_mat_h)) < ncol(com_mat_h),] 

com_mat_h[is.na(com_mat_h)] <- 0 
com_mat_h <- com_mat_h[rowSums(com_mat_h) > 0,] 

ses_h <- picante::ses.pd(samp = com_mat_h, tree = phylo_helico, null = "richness")
ses_h$pd.obs.z[is.na(ses_h$pd.obs.z)] <- 0


###Ithomiini
load("../Ithomiini/phylo_itho_correct.RData")

i_sp_stack <- readAll(brick(readRDS("../Ithomiini/All_sp_proba_stack_Jaccard.80.RDS")))
com_mat_i <- i_sp_stack@data@values

com_mat_i <- com_mat_i[,colnames(com_mat_i) %in% phylo_itho$tip.label]
row.names(com_mat_i) <- paste0("c",1:nrow(com_mat_i))
com_mat_i <- com_mat_i[rowSums(is.na(com_mat_i)) < ncol(com_mat_i) ,] 

com_mat_i[is.na(com_mat_i)] <- 0 
com_mat_i <- com_mat_i[rowSums(com_mat_i) > 0,] 

ses_i <- picante::ses.pd(samp = com_mat_i, tree = phylo_itho, null = "richness")


ses_i$pd.obs.z[is.na(ses_i$pd.obs.z)] <- 0

save(list = c("ses_i", "ses_h"), file = "../outputs/Community_structure/SESPD_by_community_HelIth.RDATA")

# load("../outputs/Community_structure/SESPD_by_community_HelIth.RDATA")
#Map Z scores
# 
# ses_i$inter <- 0
# ses_i[ses_i$pd.obs.z > 0 & ses_i$pd.obs.p <= 0.05] <- 1
# ses_i[ses_i$pd.obs.z < 0 & ses_i$pd.obs.p <= 0.05] <- -1

par(mfrow = c(1,2))
zSES_h <- Full_indices_H[[1]]
zSES_hv <- zSES_h@data@values
zSES_hv[!is.na(zSES_hv)] <- NA
zSES_hv[as.numeric(substr(row.names(ses_h), 2,8))] <- ses_h$pd.obs.z
zSES_h@data@values <- zSES_hv
 zSES_h@data@values[Full_indices_H[[1]]@data@values <= 1 & Full_indices_H[[1]]@data@values > 0 ] <- 0.0001
zSES_h@data@values[Full_indices_H[[1]]@data@values == 0 ] <- 0
plot(zSES_h)


zSES_i <- i_sp_stack[[1]]
zSES_iv <- zSES_i@data@values
zSES_iv[!is.na(zSES_iv)] <- NA
zSES_iv[as.numeric(substr(row.names(ses_i), 2,8))] <- ses_i$pd.obs.z
zSES_i@data@values <- zSES_iv

plot(zSES_i)

###PSV
# psv_h <- picante::psv(samp = com_mat_h, tree = phylo_helico,compute.var=TRUE,scale.vcv=TRUE)
# psv_i <- picante::psv(samp = com_mat_i, tree = phylo_itho,compute.var=TRUE,scale.vcv=TRUE)
# 
# par(mfrow = c(1,2))
# 
# zpsv_h <- sp_richness_H
# zpsv_hv <- zpsv_h@data@values
# zpsv_hv[!is.na(zpsv_hv)] <- NA
# zpsv_hv[as.numeric(substr(row.names(psv_h), 2,8))] <- psv_h$PSVs
# zpsv_h@data@values <- zpsv_hv
# 
# plot(zpsv_h)
# 
# zpsv_i <- i_sp_stack[[1]]
# zpsv_iv <- zpsv_i@data@values
# zpsv_iv[!is.na(zpsv_iv)] <- NA
# zpsv_iv[as.numeric(substr(row.names(psv_i), 2,8))] <- psv_i$PSVs
# zpsv_i@data@values <- zpsv_iv
# 
# plot(zpsv_i)

saveRDS(zSES_h, file = "../outputs/Indices_Maps/SESPD.raster_Jaccard.80.rds")
saveRDS(zSES_i, file = "../outputs/Indices_Maps/SESPD_Ith.raster_Jaccard.80.rds")
```


```{r}
# 3.2/ Project all maps

# Put all maps in a list

list_all_maps <- list(sp_richness, sp_mean_geo_rarity, Faith_PD, zSES_h, ring_richness_ss, ring_mean_geo_rarity_ss, weighted_mean_ring_size_ss, Andes_sp_richness, Andes_ring_mean_geo_rarity_ss, Andes_weighted_mean_ring_size_ss, ring_richness_sl, ring_mean_geo_rarity_sl, weighted_mean_ring_size_sl,  Andes_ring_mean_geo_rarity_sl, Andes_weighted_mean_ring_size_sl)
names(list_all_maps) <- c("sp_richness", "sp_mean_geo_rarity", "Faith_PD", "zSES","ring_richness_ss", "ring_mean_geo_rarity_ss", "weighted_mean_ring_size_ss"," Andes_sp_richness", "Andes_ring_mean_geo_rarity_ss", "Andes_weighted_mean_ring_size_ss", "ring_richness_sl", "ring_mean_geo_rarity_sl", "weighted_mean_ring_size_sl",  "Andes_ring_mean_geo_rarity_sl", "Andes_weighted_mean_ring_size_sl")



#Red blue palette for SS
p1 <- c(paletteer::paletteer_c("ggthemes::Red-Blue Diverging", n =50))[1:25]
p2 <- c(paletteer::paletteer_c("ggthemes::Red-Blue Diverging", n =150))[75:150]
pal_rb <- c(p1,p2)
```


```{r}
for(m in c("ss", "sl"))
{
pdf(file = paste0("../Maps/Index_maps_Mollweide_ses_",m,".pdf"), height = 12, width = 12)

internal_margins <- par()$mar
# par(mar = c(3.1, 3.5, 3.5, 2.1))
par(mar = c(3.1, 3.1, 2.7, 1.6))
par(mfrow = c(3, 3))

# A/ Species richness ####

map_indices_Mollweide(x = sp_richness,
                      main_title = "Species richness",
                      legend_title = "Species",
                      legend_title_x = -3650,
                      legend_title_y = 430,
                      # legend_breaks = seq(0, 35, 5), 
                      facet_letter = "(a)")

# B/ Species mean geographic rarity ####

# Need to add a few values to lower values to avoid being colored as null outer-range values
sp_mean_geo_rarity[(sp_mean_geo_rarity[] > 0) & (sp_mean_geo_rarity[] <= 0.005)] <- 0.005

map_indices_Mollweide(x = sp_mean_geo_rarity,
                      main_title = "Species geographic rarity",
                      legend_title = "Rarity\nindex",
                      legend_title_x = -3650,
                      legend_title_y = 870,
                      # legend_breaks = seq(0, 0.5, 0.1), 
                      facet_letter = "(b)")

# C/ Faith's PD ####

map_indices_Mollweide(x = zSES_h,
                      main_title = "Phylogenetic diversity ",
                      legend_title = "SES Z-score",
                      legend_title_x = -3650,
                      legend_title_y = 670,
                      color_palette = pal_rb,
                      legend_breaks = seq(-1, 3, 1), 
                      facet_letter = "(c)")

# D/ Mimicry richness ####
ring_richness <- get(paste0("ring_richness_",m))

map_indices_Mollweide(x = ring_richness,
                      main_title = "Mimicry richness",
                      legend_title = "Mimicry\nrings",
                      legend_title_x = -3670,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 24, 4), 
                      facet_letter = "(d)")

# E/ Mimicry ring mean geographic rarity ####
ring_mean_geo_rarity <- get(paste0("ring_mean_geo_rarity_",m))

map_indices_Mollweide(x = ring_mean_geo_rarity,
                      main_title = "Mimicry geographic rarity",
                      legend_title = "Rarity\nindex",
                      legend_title_x = -3650,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 0.44, 0.1), 
                      facet_letter = "(e)")

# F/ Mean ring size ####
weighted_mean_ring_size <- get(paste0("weighted_mean_ring_size_",m))

map_indices_Mollweide(x = weighted_mean_ring_size,
                      main_title = "Mean mimicry ring size",
                      legend_title = "Ring\nsize",
                      legend_title_x = -3750,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 3, 0.5), 
                      facet_letter = "(f)")

# G/ Species richness in Andes ####

map_indices_Andes_Mollweide_down(x = sp_richness,
                            main_title = "Species richness",
                            legend_title = "Species",
                            y_adjust_lvl = 0.0,            # Adjustment for elevation level legend
                            y_adjust_elevation = 0, 
                            # legend_breaks = seq(0, 35, 5), 
                            facet_letter = "(g)",
                            facet_letter_col = "white")

# H/ Mean species geographic rarity in Andes ####

map_indices_Andes_Mollweide_down(x = sp_mean_geo_rarity,
                                 color_palette = pal_bl_red_Mannion[-1],
                                 main_title = "Species geographic rarity",
                                 legend_title = "Rarity index",
                                 legend_title_x = -1100,
                                 legend_title_y = 350,
                                 y_adjust_lvl = 0.0,
                                 y_adjust_elevation = 0,
                                 # legend_breaks = seq(0, 0.5, 0.1), 
                                 facet_letter = "(h)",
                                 facet_letter_col = "white")

# I/ Mean ring size in Andes ####
weighted_mean_ring_size <- get(paste0("weighted_mean_ring_size_",m))


map_indices_Andes_Mollweide_down(x = weighted_mean_ring_size,
                                 main_title = "Mean mimicry ring size",
                                 legend_title = "Ring size",
                                 legend_title_x = -1200,
                                 # legend_title_y = 350,
                                 elevation = TRUE,
                                 # legend_breaks = seq(0, 3, 0.5), 
                                 facet_letter = "(i)",
                                 facet_letter_col = "white")

par(mar = internal_margins)
par(mfrow = c(1, 1))

dev.off()
}

```

```{r}
#Ithomiini SES regression
### Load maps 

for(m in c("ss", "sl"))
{
####Helico
sp_richness_H <- readRDS(file = paste0("../outputs/Indices_maps/tot.sp.richness_Jaccard.80.rds"))
# sp_mean_geo_rarity <- readRDS(file = "../outputs/Indices_maps/sp.mean.rarity_Jaccard.80.rds")       # Linear weighting
sp_mean_geo_rarity_H <- readRDS(file = "../outputs/Indices_maps/sp.mean.rarity_Leroy_Jaccard.80_contrasted.rds")   # Leroy's weighting
Faith_PD_H <- readRDS(file = "../outputs/Indices_Maps/PD.raster_Jaccard.80.rds")

load(file = paste0("../outputs/Indices_maps/ring.richness_Jaccard_",m,".80.Rdata"))
ring_richness_H<-readAll(readRDS(paste0("../outputs/Indices_maps/ring.richness_Jaccard_",m,".80.rds")))
# ring_mean_geo_rarity <- readRDS(file = "../outputs/Indices_maps/mimicry.mean.rarity_Jaccard.80.rds")  # Linear weighting
ring_mean_geo_rarity_H <- readRDS(file = paste0("../outputs/Indices_maps/ring.mean.rarity_Leroy_Jaccard_",m,".80_contrasted.rds") )   # Leroy's weighting
weighted_mean_ring_size_H <- readRDS(file = paste0("../outputs/Indices_maps/weighted_mean_ring_size_",m,".rds"))

Full_indices_H <- stack(sp_richness_H, sp_mean_geo_rarity_H, zSES_h, ring_richness_H, ring_mean_geo_rarity_H, weighted_mean_ring_size_H)
names(Full_indices_H)<-c("sp_richness", "sp_mean_geo_rarity", "zSES", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

for(i in 2:nlayers(Full_indices_H))
{
  Full_indices_H[[i]]@data@values[Full_indices_H[[i]]@data@values == 0 & is.na(Full_indices_H[[1]]@data@values) == FALSE]
}

saveRDS(Full_indices_H, paste0("../outputs/Full_Helico_indice_stack_",m,"_SES.RDS"))

}

#Itho

Full_indices_I <- readRDS( file="../Ithomiini/Full_indices_I.RDS")
names(Full_indices_I)<-c("sp_richness", "sp_mean_geo_rarity", "FaithPD", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

zSES_ib <- projectRaster(zSES_i, Full_indices_I[[1]], method = "bilinear")

Full_indices_I[["FaithPD"]] <- zSES_ib
names(Full_indices_I)<-c("sp_richness", "sp_mean_geo_rarity", "zSES", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

Full_indices_I<-projectRaster(from=Full_indices_I, crs = crs(Full_indices_H), method="bilinear")
Full_indices_I<-resample(Full_indices_I, Full_indices_H[[1]])
saveRDS(Full_indices_I, file="../Ithomiini/Full_indices_I_SES.RDS")


# Full_indices_H <- readRDS(paste0("../outputs/Full_Helico_indice_stack_",m,"_SES.RDS"))

Full_indices_I <- readRDS(file="../Ithomiini/Full_indices_I_SES.RDS")
# Full_indices_I[["zSES"]]@data@values[is.na(Full_indices_I[["zSES"]]@data@values) & !is.na(Full_indices_H[["zSES"]]@data@values)] <- 0

# Full_indices_H[["zSES"]][Full_indices_H[["sp_richness"]]  < 0.1 && Full_indices_H[["zSES"]] == 0] <- 0.0001

```

```{r}
###Itho diversity plot
# Put all maps in a list

#Red blue palette for SS
pal_rb <- c(paletteer::paletteer_c("ggthemes::Red-Blue Diverging", n =100))
```


```{r}

pdf(file = paste0("../Maps/Index_maps_Mollweide_ses_Itho.pdf"), height = 8, width = 12)

internal_margins <- par()$mar
# par(mar = c(3.1, 3.5, 3.5, 2.1))
par(mar = c(3.1, 3.1, 2.7, 1.6))
par(mfrow = c(2, 3))

# A/ Species richness ####

map_indices_Mollweide(x = Full_indices_I[[1]],
                      main_title = "Species richness",
                      legend_title = "Species",
                      legend_title_x = -3650,
                      legend_title_y = 430,
                      # legend_breaks = seq(0, 35, 5), 
                      facet_letter = "(a)")

# B/ Species mean geographic rarity ####

# Need to add a few values to lower values to avoid being colored as null outer-range values
sp_mean_geo_rarity[(sp_mean_geo_rarity[] > 0) & (sp_mean_geo_rarity[] <= 0.005)] <- 0.005

map_indices_Mollweide(x = Full_indices_I[[2]],
                      main_title = "Species geographic rarity",
                      legend_title = "Rarity\nindex",
                      legend_title_x = -3650,
                      legend_title_y = 870,
                      # legend_breaks = seq(0, 0.5, 0.1), 
                      facet_letter = "(b)")

# C/ Faith's PD ####

map_indices_Mollweide(x = Full_indices_I[[3]],
                      main_title = "Phylogenetic diversity ",
                      legend_title = "SES Z-score",
                      legend_title_x = -3650,
                      legend_title_y = 670,
                      color_palette = pal_rb,
                      # legend_breaks = seq(-3, 3, 1), 
                      facet_letter = "(c)")

# D/ Mimicry richness ####

map_indices_Mollweide(x = Full_indices_I[[4]],
                      main_title = "Mimicry richness",
                      legend_title = "Mimicry\nrings",
                      legend_title_x = -3670,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 24, 4), 
                      facet_letter = "(d)")

# E/ Mimicry ring mean geographic rarity ####

map_indices_Mollweide(x = Full_indices_I[[5]],
                      main_title = "Mimicry geographic rarity",
                      legend_title = "Rarity\nindex",
                      legend_title_x = -3650,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 0.44, 0.1), 
                      facet_letter = "(e)")

# F/ Mean ring size ####

map_indices_Mollweide(x = Full_indices_I[[6]],
                      main_title = "Mean mimicry ring size",
                      legend_title = "Ring\nsize",
                      legend_title_x = -3750,
                      legend_title_y = 670,
                      # legend_breaks = seq(0, 3, 0.5), 
                      facet_letter = "(f)")


par(mar = internal_margins)
par(mfrow = c(1, 1))

dev.off()


```



```{r}
#Mollweide map, create legend for each map
names_indices_clean_lg <- c("Species", "SES-PD\n(Z-scores)",  "Mean rarity", "Phen. groups",  "Mean size", "Mean rarity",)
names_indices_clean <- c("Species richness", "Phylogenetic diversity", "Species geographic rarity", "Phenotypic richness" , "Mean phenotypic group size", "Phenotypic geographic rarity")
names_indices <- c("sp_richness", "zSES", "sp_mean_geo_rarity", "ring_richness", "ring_mean_geo_rarity", "weighted_mean_ring_size")

#extract values


source("../functions/colmat_modified.R")
source("../functions/add_bivar_legend.R")
source("../functions/biv_map_modified.R")

for(m in c("ss", "sl"))

{
  Full_indices_H <-  readRDS(paste0("../outputs/Full_Helico_indice_stack_",m,"_SES.RDS"))
  


#Number of quantiles
nquantiles = 4

for(i in 1:length(names_indices_clean))
  {
  names <- names_indices_clean[i]
  

  #Yellow Blue, colorblind friendly
  if(names != "Phylogenetic Diversity")
  {
  col.matrix <- colmat_modif(nquantiles=4, upperleft="#f3b300", upperright="#000000", bottomleft="#f3f3f3", bottomright="#509dc2", xlab=paste0("Heliconiini ", names), ylab= paste0("Ithomiini ", names), valuesx = Full_indices_H[[i]], valuesy =  Full_indices_I[[i]])
  
    bivmap <- biv_map_modified(Full_indices_H[[i]], Full_indices_I[[i]], col.matrix, nquantiles = 4)
  }else{
    #
    col.matrix <- matrix(c("#b2b796","#b2b796","#bbd6e1","#4f9dc1",
                           "#b2b796","#b2b796","#bbd6e1","#4f9dc1",
                           "#f3dda3","#f3dda3","#f3f3f3","#346881",
                           "#f3b300","#f3b300","#a17700", "#000000"
                    ), nrow = 4, ncol = 4, byrow= T)


    # col.matrix <- matrix(c("#ADDCA1FF","#ADDCA1FF","#83CFBAFF","#509dc2",
    #                        "#ADDCA1FF","#ADDCA1FF","#83CFBAFF","#509dc2",
    #                        "#DACB64FF","#DACB64FF","#DCD4D0FF","#4DBAC3FF",
    #                        "#f3b300","#f3b300","#A6BE5AFF", "#22763FFF"
    #                 ), nrow = 4, ncol = 4, byrow= T)
    #
    bivmap <- biv_map_modified(Full_indices_H[[i]], Full_indices_I[[i]], col.matrix, nquantiles = 3, seqx = seq(-3,3, length.out = 4), seqy = seq(-3,3, length.out = 4))
    # bivmap@data@values[Full_indices_H[[1]]@data@values == 0] <-8
  }

  

  #   
  assign(paste0(names_indices[i], "_bivmap_",m), bivmap)
  
  #Add grey for 0x0 as last color
  col.matrix_v <- as.vector(col.matrix)
  col.matrix_v[ length(col.matrix) + 1] <-  "lightgrey" #"#DEE1EB"
  
  assign(paste0(names_indices[i], "_colmat"), col.matrix)
  assign(paste0(names_indices[i], "_colv"), col.matrix_v)
  
}

for(l in 1:6)
{
name_code <- names_indices[l]
col.matrix =   get(paste0(name_code, "_colmat"))

scale_decimals <- 1 #General case, not possible to have more numbers so for big number, no decimals

if(max(Full_indices_H[[l]]@data@values, na.rm= TRUE) > 9.99)
{
 scale_decimals <- 0 
}

if(name_code != "zSES")
{
  nquant <- 4
sx <- NA
sy <- NA
} else {
nquant <- 3
sx <- sy <- seq(-3,3, length.out = 4)
}


add_bivar_legend(rasterx = Full_indices_H[[l]],
                 rastery = Full_indices_I[[l]],
                 color.matrix = col.matrix,
                 nquantiles = nquant,
                 xlab = substitute(paste(bold("Heliconiini"))),
                 ylab = substitute(paste(bold("Ithomiini"))),
                 pngname = paste0("../Maps/Map_legends/",name_code,"_Bivmap_legend_",m,"_SES.png" ),
                 scaledeci = scale_decimals,
                 seqx = sx,
                 seqy= sy)
}                 
}

#Create lines for distribution borders

distrib_borders <- sp_richness_bivmap_ss
distrib_borders@data@values[distrib_borders@data@values > nquantiles^2] <- NA
distrib_borders@data@values[distrib_borders@data@values <= nquantiles^2] <- 1

distrib_borders <- boundaries(distrib_borders
                , type='outer')
distrib_borders@data@values[distrib_borders@data@values == 0] <- NA

distrib_borders <- mask(distrib_borders, sp_richness_bivmap_ss)

```


```{r}
#Mollweide map, final plot

source("../functions/custom_Bivariates_maps_mollweide.R")

for(m in c("ss", "sl"))
{

pdf(paste0("../Maps/BivarMap_",m,"_SES.pdf"), height = 8, width = 12)
par(mar = c(3.1, 3.1, 2.7, 1.6))
par(mfrow = c(2, 3))

for (l in 1:6)
{
name_code <- names_indices[l]
name_clean <- names_indices_clean[l]
name_clean_lg <- names_indices_clean_lg[l]

map <- get(paste0(name_code, "_bivmap_",m))
palette_bv <- unique(get(paste0(name_code, "_colv")))
  
 image(sea_shp,  col="#E1f1f1", #background
         xlim = c(-5457, 4953), ylim = c(-4495, 5705), axes = F,
         xlab = "", ylab = "")
# Add background, borders and graticules
  plot(bbox_sp_Mollweide, lwd = 2, border = "black", col = NA, legend=F, add = T)
  plot(grid_Mollweide_out, lty = 92, col = "grey80",legend=F,  add = T)
  par(new=TRUE)

Bivar_map_Mollweide(x = map, color_palette = palette_bv, main_title = name_clean, legend_title = name_clean_lg, facet_letter = paste0("(",letters[l],")") )
  plot(distrib_borders, col = "red", add = T, legend = F)
plot(large_bg_mask_Mollweide, lwd = 1, col = "grey20", add = T)
par(new=TRUE)

legend_biv <- readPNG(source = paste0("../Maps/Map_legends/",name_code,"_Bivmap_legend_",m,"_SES.png"))

par(new = TRUE)
rasterImage(legend_biv, xleft = -5500, ybottom = -4400, xright = -1000, ytop = 100)


}
dev.off()
}
```
