==> scripts/20_Index_bis.Rmd <==
---
title: "20_Index_bis"
author: "Eddie Pérochon"
date: "19/05/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

##### Script 20: Compute and plot extra indices maps #####

###################################################
#      Authors: Eddie Pérochon & Maël Doré        #
#      Contact: eddie.perochon@hotmail.com        #
###################################################


### Goals = 

### Generate new index maps ###

# Mean ring size
# Weighted Mean ring Size
# Max ring Size
# Quantile 25%, 50%, 75% ring size
# Mean ring phenotypic prevalence (Mean size of rings found locally. Size on the whole tribe/range)
# Mean ring phenotypic rarity (Same but phenotypic rarity of a pattern as 1 - normalized counts)
# OMUs mean phenotypic rarity (Same but weighted at OMU level)
# (Weighted) Community vulnerability as in CHazot et al., 2016

### Nice plot for extra indices in Supplementary
# Maps for supplementary with Shannon's indices, Evolutionary Distinctiveness and Maximum mimicry ring size
 
### Inputs 
   # Stack of unit 'probability' of presence from script 15
   # Stack of mimicry ring 'probability' of presence from script 18
   # Stack of mimicry ring richness of presence from script 18

###

### Outputs
    # Extra indices maps
   # Table of phenotypic prevalence and rarity
   # Nice maps of extra indices for Supplementaries
        # Shannon's species diversity
        # Shannon's mimicry diversity
        # Evolutionary Distinctiveness aas Sum of Fair-Proportions
        # Max ring Size

###

```{r}
# Clean environment
rm(list = ls())


library(raster)
library(prettymapr)
library(rangeBuilder)
library(dplyr)
library(sp)
library(rgdal)
library(stringr)
```


```{r grid and black box}
# ####Create bbo_sp
# bbox<-continent_mask
# bbox@data@values[bbox@data@values==0]<-NA
# empty<-bbox #for black box
# 
# empty[1,]<-1
# empty[,1]<-1
# empty[347,]<-1
# empty[,340]<-1
# 
# values(bbox)<-values(empty)
# 
# saveRDS(bbox, file="../input_data/Map_stuff/bbox_sp_Mollweide.rds")
# 
# 
# empty2<-projectRaster(empty, crs = CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"))
# grid<-empty2
# 
# for( i in 1:7)
# {
#    empty2[,64*i]<-1
# 
# }
# for (i in 1:6) {
#    
# 
#    empty2[62*i,]<-1
#    }
# 
#    
# grid<-empty2
# grid<-projectRaster(grid, crs = CRS("+proj=moll +lon_0=-79 +lat_0=6 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=km +no_defs"))
# 
# grid<-crop(grid, bbox)
# 
# for (i in 1:ncell(continent_mask)) {
#    if(is.na(continent_mask[i])==FALSE)
#    {
#       grid[i]<-NA
#    }
# }
# saveRDS(grid, file = "../input_data/Map_stuff/grid_Mollweide_out.rds")
```

```{r Objects}
# New color palette
pal_bl_red_Mannion <- readRDS(file = "../maps/pal_bl_red_Mannion.rds")

# Load mask for continent borders, plot border, and grid
grid_Mollweide_out <- readRDS(file = "../input_data/Biogeographic_data/grid_Mollweide_out.rds")
large_bg_mask_Mollweide <- readRDS(file = "../input_data/envData/crop_shape.RDS")
bbox_sp_Mollweide <- readRDS(file = "../input_data/Biogeographic_data/bbox_sp_Mollweide.rds")
Community_mask <- readRDS(file = "../input_data/Biogeographic_data/Community_mask.rds")
continent_mask <- readRDS(file = paste0("../input_data/envData/continent_mask.RDS"))

### Load maps
sp.richness <- readRDS(file = paste0("../outputs/Indices_maps/tot.sp.richness_Jaccard.80.rds"))
ring.richness <- readRDS(file = paste0("../outputs/Indices_maps/ring.richness_Jaccard.80.rds"))
vulnerability <- readRDS(file = paste0("../outputs/Indices_maps/vulnerability_Jaccard.80.rds"))

# Load mimicry ring richness stack
# load(file="../outputs/Indices_stacks/All_ring_rich_stack_Jaccard.80.RData")
# ring_rich_stack<-All_ring_rich_stack_Jaccard.80

# Load mimicry ring proba stack
ring_proba_stack <- readRDS(file = paste0("../outputs/Indices_stacks/All_ring_proba_stack_Jaccard.80.RData"))

```

```{r Functions}
# 2.2/ Plotting function ####

# Function to map indices
{
  map_indices_Mollweide <- function(x,                                    # Raster to map
                                    color_palette = pal_bl_red_Mannion,   # Color palette
                                    main_title,                           # Main title
                                    main_title_cex = 1.5,                 # Main title size
                                    
                                    xlim = c(-5457, 4953),   # Limit of plot on x-axis (Longitude)
                                    ylim = c(-4495, 5705),    # Limit of plot on y-axis (Latitude)
                                    axis_cex = 1.4,             # Axes size
                                    
                                    xlab = "",                # X-axis label
                                    ylab = "",                # Y-axis label
                                    x_axis_breaks = c(-3930, -2170, -420, 1500, 3050),            # X-axis tick breaks
                                    y_axis_breaks = c(-3650, -2450, -1220, 0, 1230, 2445),        # Y-axis tick breaks
                                    x_axis_labels = c("120°E", "100°E", "80°E", "60°E", "40°E"),      # X-axis tick labels
                                    y_axis_labels = c("30°S", "20°S", "10°S", "0°", "10°N", "20°N"),  # Y-axis tick labels
                                    
                                    legend_title,             # Legend title
                                    legend_title_cex = 1.3,   # Legend title size
                                    legend_title_x = -3550,   # Legend title x position
                                    legend_title_y = 430,     # Legend title y position
                                    legend_cex = 1.4,         # Legend size
                                    legend_breaks,            # Legend tick positions
                                    legend_location = c(-4100, -3800, -3950, 0),  # Legend position
                                    
                                    scale_bar_position = c(-2600, -4000),  # Scale bar position
                                    
                                    arrow_scale = 0.55,           # North arrow size
                                    arrow_padin = c(0.15, 0.15),  # North arrow position adjustement
                                    
                                    facet_letter = "",                  # Small case letter for facet
                                    facet_letter_col = "black",         # Color of case letter for facet
                                    facet_letter_cex = 2.2,             # Size of small case letter for facet
                                    facet_letter_inset = c(0, -0.008))  # Position adjustment of small case letter for facet
  
  {
    # Plot raster background without axis
    image(x, col = color_palette,
          xlim = xlim, ylim = ylim, axes = F,
          xlab = xlab, ylab = ylab)
    title(main = main_title, cex.main = main_title_cex, line = 1)
    
    # Generate axes with manual positioning of ticks
    axis(side = 1, at = x_axis_breaks, labels = x_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0, padj = 0.5)
    axis(side = 2, at = y_axis_breaks, labels = y_axis_labels, cex.axis = axis_cex, lwd = 0.2, lwd.ticks = 1, gap.axis = 0)
    
    # Add background, borders and graticules
    plot(large_bg_mask_Mollweide, lwd = 1, border = "grey20", col = "aliceblue", add = T)
    plot(grid_Mollweide_out, lty = 92, col = "grey80", add = T)
    plot(bbox_sp_Mollweide, lwd = 2, border = "black", col = NA, add = T)

    
    # Add scale bar in legend
    scalebar(d = 2000, type = "line", lwd = 4, divs = 4, xy = scale_bar_position, label = c("", "2000 km", ""), adj = c(0.5, -0.8), font = 2, cex = 1.2)
    prettymapr::addnortharrow(scale = arrow_scale, padin = arrow_padin, text.col = "#00000000")
    rangeBuilder::addRasterLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
    rangeBuilder::addRasterLegend(x, locs = legend_breaks, cex.axis = legend_cex, ramp = color_palette, ncolors = 200, border = T, location = legend_location)
    graphics::text(x = legend_title_x, y = legend_title_y, font = 2, cex = legend_title_cex, label = legend_title)
    
    # Add facet letter
    legend(legend = facet_letter, x = "bottomright", bty = "n", text.col = facet_letter_col,
           text.font = 2, cex = facet_letter_cex, inset = facet_letter_inset)
    
  }
}

# map_indices_Mollweide(x = sp.richness_Mollweide,
#                       main_title = "Species richness",
#                       legend_title = "Species",
#                       legend_breaks = seq(0, 120, 20), 
#                       facet_letter = "a")

# 2.3/ Add continental null values to raster ####

add_continental_null_values <- function(x)
{
  x@data@values[str_detect(x@data@values, "NaN" )]<-NA
  y <- continent_mask  # Create final new raster from continental mask
  for (i in 1:ncell(x)) {
     if(is.na(x[i])==TRUE && is.na(y[i])==FALSE)
     {x[i]<-y[i]} 
  }
# Add initial raster values
  
  return(x)
}

# 2.4/ Contrasting raster ####
contrasting_raster <- function(x, zmin, zmax)
{
  Community_mask <- readRDS(file = "../input_data/Map_stuff/Community_mask.rds")
  continent_mask <- readRDS(file = paste0("../input_data/envData/continent_mask.rds"))
  
  x[x[] <= zmin] <- zmin  # Fix low values
  x[x[] >= zmax] <- zmax  # Fix high values
  
  x <- mask(x, mask = Community_mask)  # Cut out values that are outside Ithomiini range
  
  y <- continent_mask + zmin  # Create final new raster from continental mask with baseline = zmin
  y[!is.na(x[])] <- x[!is.na(x[])]  # Add initial raster values

  return(y)
}

```

```{r Mean ring size}
### 3/ Compute mean ring size ####

# 3.1/ Compute index

for(m in c("ss","sl"))
{

  ring.richness <- readRDS(file = paste0("../outputs/Indices_maps/ring.richness_Jaccard_",m,".80.rds"))
  
mean.ring.size <- sp.richness/ring.richness
mean.ring.size<-add_continental_null_values(mean.ring.size)
save(mean.ring.size, file = paste0("../outputs/Indices_maps/mean.ring.size_",m,".RData"))
saveRDS(mean.ring.size, file = paste0("../outputs/Indices_maps/mean.ring.size_",m,".RDS"))

# 3.2/ Project into Mollweide and save results

# # 3.3/ Plot 
# map_indices_Mollweide(x = mean.ring.size,
#                       main_title = "Mean ring size",
#                       legend_title = "Ring size",
#                       legend_title_y = 530,
#                       legend_breaks = seq(1, 2.5, 0.5), 
#  
#                                            facet_letter = "a")
}

```

```{r Weigthed mean ring size}

for(m in c("ss", "sl"))
{
### 4/ Weighted mean ring size ####

# (weights = proba of presence of the ring)

# Load mimicry ring richness stack
All_ring_rich_stack_Jaccard.80 <- readRDS(file=paste0("../outputs/Indices_stacks/All_ring_rich_stack_Jaccard_",m,".80.RDS"))
ring_rich_stack<-All_ring_rich_stack_Jaccard.80

# Load mimicry ring proba stack
ring_proba_stack <- readRDS(file = paste0("../outputs/Indices_stacks/All_ring_proba_stack_Jaccard_",m,".80.RDS"))

# 4.1/ Compute index

# ?weighted.mean
# test <- overlay(x = ring_rich_stack, y = ring_proba_stack, fun = weighted.mean, na.rm = T, unstack=F)

ring_rich_brick <- brick(ring_rich_stack)
ring_proba_brick <- brick(ring_proba_stack)
# # 
# View(ring_rich_brick[])

weighted_mean_ring_size <- vector()
for (i in 1:nrow(ring_rich_brick[]))  # Loop between communities/pixels
{
  # i <- 55000
  rich_vector <- ring_rich_brick[i]   # Extract ring richness
  proba_vector <- ring_proba_brick[i] # Extract ring proba of presence
  
  # Compute mean weighted by proba. of presence
  weighted_mean_ring_size[i] <- weighted.mean(x = rich_vector, w = proba_vector)
  if(i%%1000==0){print(i)}
}
 
weighted_mean_ring_size_mask<-continent_mask
weighted_mean_ring_size_mask@data@values<-weighted_mean_ring_size
weighted_mean_ring_size_mask <- add_continental_null_values(weighted_mean_ring_size_mask)
save(weighted_mean_ring_size_mask, file = paste0("../outputs/Indices_maps/weighted_mean_ring_size_",m,".RData"))
saveRDS(weighted_mean_ring_size_mask, file = paste0("../outputs/Indices_maps/weighted_mean_ring_size_",m,".rds"))

# 4.2/ Project into Mollweide and save results

# 4.3/ Plot 
# map_indices_Mollweide(x = weighted_mean_ring_size_mask,
#                       main_title = "Weighted mean ring size",
#                       legend_title = "Ring size",
#                       legend_title_y = 530,
#                       legend_breaks = seq(0, 3, 0.5), 
#                       facet_letter = "a")

}

```

```{r Max ring size}
### 5/ Max ring size ####

# 5.1/ Compute index
max_ring_size <- calc(ring_rich_stack, fun = max)*1
save(max_ring_size, file = "../outputs/Indices_maps/max_ring_size.RData")
saveRDS(max_ring_size, file = "../outputs/Indices_maps/max_ring_size.rds")



# 5.3/ Plot 
map_indices_Mollweide(x = max_ring_size,
                      main_title = "Max ring size",
                      legend_title = "Ring size",
                      legend_title_y = 530,
                      legend_breaks = seq(0, 11.5, 3), 
                      facet_letter = "a")

```

```{r}
### 7/ Mean ring phenotypic rarity and prevalence ####

# Normalized_function

normalized_0_1 <- function(x, ...)
{
  if(max(x) == min(x)){
    y <- x/length(x)
  } else {
    y <- (x-min(x))/(max(x)-min(x))
  }
  return(y)
}


# 7.1/ Compute ring phenotypic rarity based on counts of OMUs per ring

# Extract order of ring layers in stack
ring_ordered <- data.frame(names(ring_proba_stack)) ; names(ring_ordered) <- "ring"

load(file = paste0("../input_data/occurences_df/list_occ_after_clipping_rs.RData"))

list_OMU<-data.frame(OMU=unique(list_occ$OMU))
for(i in 1:nrow(list_OMU))
{list_OMU$Mimicry_ring[i]<-list_occ$Mimicry_ring[list_occ$OMU==list_OMU$OMU[i]]}
list_OMU$Mimicry_ring<-str_replace_all(list_OMU$Mimicry_ring, " ", ".")
list_OMU$Mimicry_ring<-str_replace_all(list_OMU$Mimicry_ring, "-", ".")

N_OMUs_per_ring <- list_OMU %>% 
  group_by(Mimicry_ring) %>% 
  summarize(counts = n()) %>% 
  ungroup() %>% 
  mutate(normed_commonness = normalized_0_1(counts)) %>% 
  mutate(phenotypic_rarity = 1 - normed_commonness) %>% 
  left_join(x = ring_ordered, y = ., by = c("ring" = "Mimicry_ring"))

save(N_OMUs_per_ring, file = paste0("../outputs/N_OMUs_per_ring.RData"))

load(file = paste0("../outputs/N_OMUs_per_ring.RData"))

hist(N_OMUs_per_ring$counts)

# Ideally, ring phenotypic rarity should be evaluated in the phenospace weighted by occurrences in the phylogeny. 
# Also phylogeny is not accounted. A way to account for phylogeny would be to count independent appearance of the pattern instead of all occurrences

# 7.2/ Compute indices

# Weighted mean of phenotypic rarity

# Weights = probability of presence of the ring
# Values for rarity based on size of the ring in the whole tribe => 1 - normalized counts of OMUs per ring

mean_ring_phenotypic_rarity <- calc(x = ring_proba_stack, 
                                    fun = function(x) {weighted.mean(x = N_OMUs_per_ring$phenotypic_rarity, w = x)})*1

# Weighted mean of ring prevalence (size in the whole tribe)
mean_ring_prevalence <- calc(x = ring_proba_stack, 
                                 fun = function(x) {weighted.mean(x = N_OMUs_per_ring$counts, w = x)})*1

# Add null continental values
mean_ring_phenotypic_rarity <- add_continental_null_values(mean_ring_phenotypic_rarity)
mean_ring_prevalence <- add_continental_null_values(mean_ring_prevalence)

# Save
save(mean_ring_phenotypic_rarity, file = "../outputs/Indices_maps/mean_ring_phenotypic_rarity.RData")
save(mean_ring_prevalence, file = "../outputs/Indices_maps/mean_ring_prevalence.RData")

hist(mean_ring_phenotypic_rarity)
hist(mean_ring_prevalence)


# 7.4/ Plot
map_indices_Mollweide(x = mean_ring_phenotypic_rarity,
                      main_title = "Mean ring phenotypic rarity",
                      legend_title = "Phenotypic\nrarity",
                      legend_title_y = 730,
                      legend_breaks = seq(0, 0.8, 0.1), 
                      facet_letter = "a")

map_indices_Mollweide(x = mean_ring_prevalence,
                      main_title = "Mean ring prevalence",
                      legend_title = "Ring size",
                      legend_title_y = 530,
                      legend_breaks = seq(0, 8, 2), 
                      facet_letter = "a")

# 7.5/ Contrasted plot

# Weighted mean phenotypic rarity
hist(mean_ring_phenotypic_rarity) # 0.4 to 0.8

mean_ring_phenotypic_rarity_contrasted <- contrasting_raster(x = mean_ring_phenotypic_rarity, zmin = 0.6, zmax = 0.83)

map_indices_Mollweide(x = mean_ring_phenotypic_rarity_contrasted,
                      main_title = "Mean ring phenotypic rarity",
                      legend_title = "Phenotypic\nrarity",
                      legend_title_y = 730,
                      legend_breaks = seq(0.6, 0.83, 0.1), 
                      facet_letter = "a")

# Mean ring prevalence
hist(mean_ring_prevalence) # 25 to 55

mean_ring_prevalence_contrasted <- contrasting_raster(x = mean_ring_prevalence, zmin = 4, zmax = 8)
map_indices_Mollweide(x = mean_ring_prevalence_contrasted,
                      main_title = "Mean ring prevalence",
                      legend_title = "Ring size",
                      legend_title_y = 530,
                      legend_breaks = seq(4, 8, 0.5), 
                      facet_letter = "a")
```

```{r}


### 8/ OMUs mean phenotypic rarity ####
# Load stack of OMU proba
OMU_proba_stack<-continent_mask
for(i in 1:nrow(list_OMU))
{
  unit<-list_OMU$OMU[i]
  load(paste0("../outputs/By_OMU/",unit,"/cont_Jaccard.80_",unit,".RData"))
  OMU_proba_stack<-addLayer(OMU_proba_stack, sp.cont_Jaccard.80)
}
  OMU_proba_stack<-dropLayer(OMU_proba_stack, 1)
  names(OMU_proba_stack)<-list_OMU$OMU
  
saveRDS(OMU_proba_stack,file = paste0("../outputs/Indices_stacks/All_OMU_stack_Jaccard.80.rds") )  
# 8.1/ Load stuff


OMU_proba_stack <- readRDS(file = paste0("../outputs/Indices_stacks/All_OMU_stack_Jaccard.80.rds"))

# Load Summary table for OMUs
load(file = paste0("../input_data/occurences_df/list_occ_after_clipping_rs.RData"))
# Load summary table for rings
load(file = paste0("../outputs/N_OMUs_per_ring.RData"))

list_OMU <- list_OMU %>% 
  left_join(x = ., y = N_OMUs_per_ring[, c("ring", "phenotypic_rarity")], by = c("Mimicry_ring" = "ring"))

# 8.2/ Compute indices

# Weighted mean of phenotypic rarity

# Weights = probability of presence of the OMUs
# Values for rarity based on size of the ring in the whole tribe => 1 - normalized counts of OMUs per ring
list_OMU$OMU<-str_replace_all(list_OMU$OMU, " ", ".")
list_OMU$OMU<-str_replace_all(list_OMU$OMU, "-", ".")

identical(list_OMU$OMU, names(OMU_proba_stack))

mean_OMU_phenotypic_rarity <- calc(x = OMU_proba_stack, 
                                    fun = function(x) {weighted.mean(x = list_OMU$phenotypic_rarity, w = x)})*1

# Add null continental values
mean_OMU_phenotypic_rarity <- add_continental_null_values(mean_OMU_phenotypic_rarity)

save(list_OMU, file="../input_data/occurences_df/list_OMU.RData")

# Save
save(mean_OMU_phenotypic_rarity, file = "../outputs/Indices_maps/mean_OMU_phenotypic_rarity.RData")


# 8.4/ Plot
map_indices_Mollweide(x = mean_OMU_phenotypic_rarity,
                      main_title = "Mean OMU phenotypic rarity",
                      legend_title = "Phenotypic\nrarity",
                      legend_title_y = 730,
                      legend_breaks = seq(0, 0.8, 0.2), 
                      facet_letter = "a")

# 8.5/ Contrasted plot

# Weighted mean phenotypic rarity
hist(mean_OMU_phenotypic_rarity) # 0.4 to 0.82

mean_OMU_phenotypic_rarity_contrasted <- contrasting_raster(x = mean_OMU_phenotypic_rarity, zmin = 0.4, zmax = 0.82)


map_indices_Mollweide(x = mean_OMU_phenotypic_rarity_contrasted,
                      main_title = "Mean OMU phenotypic rarity",
                      legend_title = "Phenotypic\nrarity",
                      legend_title_y = 730,
                      legend_breaks = seq(0.4, 0.7, 0.1), 
                      facet_letter = "a")
```

```{r}


### 9/ Community vulnerability ####

# 9.1/ Compute unweighted version (assuming all ring with N > 0.5 are present)

# Function to compute community vulnerability as the sum of rings vulnerability approximated as the inverse of the local richness of each mimicry ring
# A ring with only one species as a vulnerability of 1/1 = 1. A ring of 4 species as a vulnerability of 1/4 = 0.25.
# Final vulnerability is standardized by total ring richness such as a community mean vulnerability

vulnerability = function(x, na.rm) {
  x <- round(x, digits = 0) # Need to round ring richness to avoid inflation of value due to numerous rings with richness values close to 0 but not null
  if (sum(x, na.rm = T) > 0) { # Computed only if local mimicry richness >= 1 once rounded
    x <- x[x>0] # Remove all 0 values to avoid error with 1/0
    V <- sum(1/x, na.rm = T) # Compute non-standardized vulnerability = sum of mimicry ring vulnerability.
    V <- V/length(x) # Standardization by local mimicry ring richness
  }else{
    V <- NA
  }
  return(V) # Output
}

vulnerability <- calc(x = ring_rich_stack, fun = vulnerability)*1

vulnerability <- add_continental_null_values(vulnerability)
save(vulnerability, file = "../outputs/Indices_maps/vulnerability.RData")

# 9.2/ Compute weighted version (weighted by ring probability of presence)

ring_proba_brick <- readAll(brick(ring_proba_stack))
ring_rich_brick <- readAll(brick(ring_rich_stack))
ring_rich_brick_rounded <- round(ring_rich_brick, digits = 0)
table(ring_rich_brick_rounded@data@values)

weighted_vulnerability <- vector()
for (i in 1:nrow(ring_rich_brick_rounded[]))  # Loop between communities/pixels
{
  # i <- 25
  
  rich_rounded_vector <- ring_rich_brick_rounded[i]   # Extract ring richness
  if (sum(rich_rounded_vector, na.rm = T) > 0) { # Computed only if local mimicry richness >= 1 once rounded
    rich_cleaned_vector <- rich_rounded_vector[rich_rounded_vector > 0] # Remove all 0 values to avoid error with 1/0
    ring_vulnerability <- 1/rich_cleaned_vector # Compute non-standardized vulnerabilities = inverse of mimicry ring richness
    
    proba_vector <- ring_proba_brick[i] # Extract ring proba of presence
    proba_vector <- proba_vector[rich_rounded_vector > 0] # Remove all ring with richness = 0
    
    # Compute mean weighted by proba. of presence
    weighted_vulnerability[i] <- weighted.mean(x = ring_vulnerability, w = proba_vector)
    
  }else{
    weighted_vulnerability[i] <- NA
  }
if (i%%1000==0){print(i)}
}
weighted_vulnerability_mask<-continent_mask
weighted_vulnerability_mask@data@values<-weighted_vulnerability
weighted_vulnerability_mask <- add_continental_null_values(weighted_vulnerability_mask)
save(weighted_vulnerability_mask, file = "../outputs/Indices_maps/weighted_vulnerability.RData")

# 9.3/ Project into Mollweide and save results


# 9.4/ Plot
map_indices_Mollweide(x = vulnerability,
                      main_title = "Community vulnerability",
                      legend_title = "Vulnerability",
                      legend_title_y = 530,
                      legend_breaks = seq(0, 1, 0.2), 
                      facet_letter = "a")

map_indices_Mollweide(x = weighted_vulnerability_mask,
                      main_title = "Community weighted vulnerability",
                      legend_title = "Vulnerability",
                      legend_title_y = 530,
                      legend_breaks = seq(0, 1, 0.2), 
                      facet_letter = "a")


# 9.5/ Contrasted plot

# Weighted mean phenotypic rarity
hist(vulnerability) # 0.4 to 1.0

vulnerability_contrasted <- contrasting_raster(x = vulnerability, zmin = 0.4, zmax = 1.0)


map_indices_Mollweide(x = vulnerability_contrasted,
                      main_title = "Community vulnerability",
                      legend_title = "Vulnerability",
                      legend_title_y = 530,
                      legend_breaks = seq(0.4, 1, 0.2), 
                      facet_letter = "a")

hist(weighted_vulnerability) # 0.4 to 1.0

weighted_vulnerability_contrasted <- contrasting_raster(x = weighted_vulnerability_mask, zmin = 0.4, zmax = 1.0)


map_indices_Mollweide(x = weighted_vulnerability_contrasted,
                      main_title = "Community weighted vulnerability",
                      legend_title = "Vulnerability",
                      legend_title_y = 530,
                      legend_breaks = seq(0.4, 1, 0.2), 
                      facet_letter = "a")

```

```{r}
# 11/ Plot for Supplementary with 4 indices ####

# 11.1/ Load stuff ###

sp.diversity <- readRDS(file = paste0("../outputs/Indices_maps/sp.diversity.compatible_Jost_Jaccard.80.rds"))
ring.diversity <- readRDS(file = paste0("../outputs/Indices_maps/ring.diversity.compatible_Jost_Jaccard.80.rds"))
ED <- readRDS(file = "../outputs/Indices_Maps/FP.raster_Jaccard.80.rds")
max_ring_size <- readRDS(file = "../outputs/Indices_maps/max_ring_size.rds")



load(file = "../input_data/Map_stuff/country_borders.RData")
country_borders <- as(country_borders, "Spatial")

# 11.3/ Plot ###

# A/ Shannon's species diversity
# B/ Shannon's mimicry diversity
# C/ Evolutionary Distinctiveness
# D/ Max ring size

pdf(file = paste0("../supplementaries/Bonus_index_maps.pdf"), height = 10, width = 10)

internal_margins <- par()$mar
par(mar = c(3.1, 3.0, 2.7, 1.5))
par(mfrow = c(2, 2))

# A/ Shannon's species diversity
map_indices_Mollweide(x = sp.diversity,
                      main_title = "Shannon's species diversity",
                      legend_title = "Effective\nspecies\nrichness",
                      legend_title_y = 1050,
                      legend_title_x = -3350,
                      legend_breaks = seq(0, 33, 5), 
                      facet_letter = "(a)")

# B/ Shannon's mimicry diversity
map_indices_Mollweide(x = ring.diversity,
                      main_title = "Shannon's mimicry diversity",
                      legend_title = "Effective\nmimicry\nrichness",
                      legend_title_y = 1050,
                      legend_title_x = -3350,
                      legend_breaks = seq(0, 18, 3), 
                      facet_letter = "(b)")

# C/ Evolutionary Distinctiveness
map_indices_Mollweide(x = ED,
                      main_title = "Evolutionary Distinctiveness",
                      legend_title = "Evolutionary\nTime (My)",
                      legend_title_y = 730,
                      legend_title_x = -2950,
                      legend_breaks = seq(1, 235, 40), 
                      facet_letter = "(c)")

# D/ Max ring size
map_indices_Mollweide(x = max_ring_size,
                      main_title = "Maximum mimicry ring size",
                      legend_title = "Ring\nsize",
                      legend_title_y = 730,
                      legend_title_x = -3650,
                      legend_breaks = seq(0, 11, 2), 
                      facet_letter = "(d)")

par(mar = internal_margins)
par(mfrow = c(1, 1))

dev.off()

```
```{r Subspecies richness}
###### 12/ subSpecies richness ######

### 1.1/ Species Stack generation ####
All_sp_proba_stack_Jaccard.80 <- All_sp_proba_stack_Jaccard.95 <- All_sp_proba_stack_TSS.80 <- All_sp_proba_stack_TSS.95  <- stack(continent_mask) # 1e temp layer used to initiate stack, to remove afterwards
# i <- 1

p<-progress_estimated(nrow(list_occ), min_time=10)
for (i in 1:nrow(list_occ)) # By sp
{ 
  p$pause(0.1)$tick()$print()
  # Load sp name
  sp <- as.character(list_occ$biomodTag[i])

if(list_occ$initial_model_type[i]!="rasterized") 
{
  # Load Sp continuous probability map 
  sp.cont_Jaccard.80 <- readRDS(file = paste0("../outputs/By_unit/",sp,"/all_EM_Jaccard_buffer_80.RDS"))[[1]]
}

if(list_occ$initial_model_type[i]=="rasterized") 
{
  # Load Sp continuous probability map 
    load(file = paste0("../outputs/By_unit/",sp,"/raster_stack_cropped.RData"))
  sp.cont_Jaccard.80 <- Select_raster_stack_cropped
}

    
  # Build stack
  All_sp_proba_stack_Jaccard.80 <- addLayer(All_sp_proba_stack_Jaccard.80, sp.cont_Jaccard.80)
  

}

# Drop the useless first layer used to initiate the stack
All_sp_proba_stack_Jaccard.80 <- dropLayer(All_sp_proba_stack_Jaccard.80, i = 1)

# Name layers with sp names
names(All_sp_proba_stack_Jaccard.80) <- list_occ$biomodTag

# nlayers(All_sp_proba_stack_Jaccard.95) # 439 subspecies in the final stack

# Save stacks
save(All_sp_proba_stack_Jaccard.80, file = paste0("../outputs/Indices_stacks/All_ssp_proba_stack_Jaccard.80.RData"), version = "2")
saveRDS(All_sp_proba_stack_Jaccard.80, file = paste0("../outputs/Indices_stacks/All_ssp_proba_stack_Jaccard.80.rds"), version = "2")


### Load directly the complete stack
All_sp_proba_stack_Jaccard.80 <- readRDS(file = paste0("../outputs/Indices_stacks/All_ssp_proba_stack_Jaccard.80.rds"))


### 1.2/ Index computation ####
tot.ssp.richness_Jaccard.80 <- readAll(calc(All_sp_proba_stack_Jaccard.80, fun = sum))


# Save
save(tot.ssp.richness_Jaccard.80, file = paste0("../outputs/Indices_maps/tot.ssp.richness_Jaccard.80.RData"), version = "2")
saveRDS(tot.ssp.richness_Jaccard.80, file = paste0("../outputs/Indices_maps/tot.ssp.richness_Jaccard.80.rds"), version = "2")


### Load directly the final Species richness layer
tot.ssp.richness_Jaccard.80 <- readRDS(file = paste0("../outputs/Indices_maps/tot.ssp.richness_Jaccard.80.rds"))

```
```{r plot SSR}
### 1.3/ Plot Species richness ####

### Individual plots

# Jaccard.80
pdf(file = paste0("../Maps/Indices_maps/Jaccard.80/tot.ssp.richness_Jaccard.80.pdf"), height = 6.5, width = 6.5)
internal_margins <- par()$mar
par(mar = c(3.1, 3.5,3.5,2.1))
image(tot.ssp.richness_Jaccard.80, col = pal_bl_red_Mannion, main = paste0("Sub-species richness \nJaccard.80"), 
      cex.axis = 1.4, cex.main = 1.4, axis.args=list(cex.axis=1.4),
      ylab = "", xlab = "",
      legend.args=list(text="          Sub-species", cex=1.2, line = 1, font = 2), 
      legend  = F )
plot(crop_shape, lwd = 1.2, border = "grey20", add = T)
addRasterLegend(tot.ssp.richness_Jaccard.80, locs = seq(0, 60, 5), cex.axis = 1.2, ramp = pal_bl_red_Mannion, ncolors = 200, border = T, location = c(-5000, -4750, -4000, -500))
graphics::text(x = -4300, y = 0, font = 2, cex = 1.2, label = "Sub-species")
par(mar = internal_margins)
dev.off()
```

```{r weighted Subspecies richness}
###### 13/  weighted subpecies richness ######

### Load directly the final Species richness layer
tot.ssp.richness_Jaccard.80 <- readRDS(file = paste0("../outputs/Indices_maps/tot.ssp.richness_Jaccard.80.rds"))
tot.sp.richness_Jaccard.80<-readRDS(file = paste0("../outputs/Indices_maps/tot.sp.richness_Jaccard.80.rds"))

weighted.ssp.richness<-tot.ssp.richness_Jaccard.80/tot.sp.richness_Jaccard.80

for(i in 1:ncell(weighted.ssp.richness))
{
 if(is.na(weighted.ssp.richness@data@values[i])==FALSE)
 {
if(weighted.ssp.richness@data@values[i]<=1) 
{weighted.ssp.richness@data@values[i]<-1}
 }
}

```
```{r plot WSSR}
### 1.3/ Plot Species richness ####

### Individual plots

# Jaccard.80
pdf(file = paste0("../Maps/Indices_maps/Jaccard.80/weighted.ssp.richness_Jaccard.80.pdf"), height = 6.5, width = 6.5)
internal_margins <- par()$mar
par(mar = c(3.1, 3.5,3.5,2.1))
image(weighted.ssp.richness, col = pal_bl_red_Mannion, main = paste0("Weighted Sub-species richness \nJaccard.80"), 
      cex.axis = 1.4, cex.main = 1.4, axis.args=list(cex.axis=1.4),
      ylab = "", xlab = "",
      legend.args=list(text="          Sub-species", cex=1.2, line = 1, font = 2), 
      legend  = F )

plot(crop_shape, lwd = 1.2, border = "grey20", add = T)
addRasterLegend(weighted.ssp.richness, locs = seq(1, 2.25, 0.25), cex.axis = 1.2, ramp = pal_bl_red_Mannion, ncolors = 200, border = T, location = c(-5000, -4750, -4000, -500))
graphics::text(x = -4300, y = 0, font = 2, cex = 1.2, label = "Sub-species per species")
par(mar = internal_margins)
dev.off()
```
```{r}
####Incertitude
load("../input_data/occurences_df/list_occ_after_clipping_rs.RData")

modeled_unit<-list_occ[list_occ$initial_model_type!="rasterized",]
IQRstack<-continent_mask
for(i in 1:nrow(modeled_unit))
{
  unit<-modeled_unit$biomodTag[i]
load(paste0("../outputs/By_unit/",unit,"/all_EM_Jaccard_buffer_80.RData"))
IQR_layer<-all_EM_Jaccard_buffer_80[[2]]
IQRstack<-addLayer(IQRstack, IQR_layer)
if(i%%10==0){print(i)}
}
IQRstack<-dropLayer(IQRstack, 1)
names(IQRstack)<-modeled_unit$biomodTag
mean_IQR<-raster::calc(readAll(IQRstack), fun=mean, na.rm=T)



```

```{r plot IQR mean}
sea_shp<-readRDS(file="../input_data/Map_stuff/sea_shp.RDS")
grid_Mollweide_out <- readRDS(file = "../input_data/Map_stuff/grid_Mollweide_out.rds")
large_bg_mask_Mollweide <- readRDS(file = "../input_data/envData/crop_shape.RDS")
bbox_sp_Mollweide <- readRDS(file = "../input_data/Map_stuff/bbox_sp_Mollweide.rds")


### Individual plots
pal_green_red<-readRDS("../Maps/pal_green_red.rds")
# Jaccard.80
pdf(file = paste0("../Maps/Indices_maps/Mean_IQR.pdf"), height = 6.5, width = 6.5)
internal_margins <- par()$mar
par(mar = c(3.1, 3.5,3.5,2.1))
image(mean_IQR, col = pal_green_red, main = paste0("Mean Interquartile range for all units"), 
      cex.axis = 1.4, cex.main = 1.4, axis.args=list(cex.axis=1.4),
      ylab = "", xlab = "",
      legend.args=list(text="          Sub-species", cex=1.2, line = 1, font = 2), 
      legend  = F )

plot(large_bg_mask_Mollweide, lwd = 1.2, border = "grey20", add = T)
plot(sea_shp, legend=F, col="#E1f1f1", add=T )
plot(large_bg_mask_Mollweide, lwd = 1, col = "grey20", add = T)
plot(grid_Mollweide_out, lty = 92, col = "grey80",legend=F,  add = T)
plot(bbox_sp_Mollweide, lwd = 2, border = "black", col = NA, legend=F, add = T)
addRasterLegend(mean_IQR, locs = seq(from=0, to=0.05, by=0.005), cex.axis = 1.2, ramp = pal_green_red, ncolors = 200, border = T, location = c(-5000, -4750, -4000, -500))
graphics::text(x = -4300, y = 0, font = 2, cex = 1.2, label = "IQR")
par(mar = internal_margins)
dev.off()
